% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available-module-functions.R
\name{available_ppc}
\alias{available_ppc}
\alias{available_mcmc}
\title{Get or view the names of available plotting functions}
\usage{
available_ppc(pattern, fixed = FALSE, invert = FALSE)

available_mcmc(pattern, fixed = FALSE, invert = FALSE)
}
\arguments{
\item{pattern, fixed, invert}{Passed to \code{\link[base:grep]{base::grep()}}.}
}
\value{
A possibly empty character vector of function names with several
additional attributes (for use by a custom print method). If \code{pattern}
is missing then the returned object contains the names of all available
plotting functions in the \link{MCMC} or \link{PPC} module, depending on
which function is called. If \code{pattern} is specified then a subset of
function names is returned.
}
\description{
Get or view the names of available plotting functions
}
\examples{
available_mcmc()
available_mcmc("nuts")
available_mcmc("rhat|neff")
available_ppc("grouped")
available_ppc("grouped", invert = TRUE)

}
