% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesplot-package.R
\docType{package}
\name{bayesplot-package}
\alias{bayesplot}
\alias{bayesplot-package}
\title{Plotting for Bayesian Models}
\description{
Plotting for Bayesian Models
}
\section{Plotting functionality}{

\itemize{
  \item \strong{\link[=MCMC-overview]{MCMC (general)}}:
  Visualizations of Markov chain Monte Carlo (MCMC) simulations generated
  by any MCMC algorithm.
  \item \strong{\link[=MCMC-overview]{MCMC (algorithm-specific)}}:
  Visualizations designed for use with particular MCMC algorithms. In future
  releases these will include HMC, HMC + NUTS, Gibbs, random walk Metropolis,
  and other variants of MCMC.
  \item \strong{\link[=Inference-overview]{Inference/Prediction}}
  Plots designed to assist with inference and prediction.
  \item \strong{\link[=PPC-overview]{PPC}}:
  Graphical posterior predictive checks (PPC).
}
}

