% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesplot_grid.R
\name{bayesplot_grid}
\alias{bayesplot_grid}
\title{Compare plots}
\usage{
bayesplot_grid(..., plots = list(), grid_args = list(), xlim = NULL,
  ylim = NULL)
}
\arguments{
\item{...}{One or more ggplot objects.}

\item{plots}{A list of ggplot objects. Can be used as an alternative to
specifying plot objects via \code{...}.}

\item{grid_args}{An optional named list of arguments to pass to
\code{\link[gridExtra]{arrangeGrob}} (\code{nrow}, \code{ncol},
\code{widths}, etc.).}

\item{xlim, ylim}{Optionally, numeric vectors of length 2 specifying lower and
upper limits for the axes that will be shared across all plots.}
}
\value{
An object of class "bayesplot_grid", which has a \code{plot} method.
}
\description{
The \code{bayesplot_grid} function makes it simple to juxtapose plots
using common \eqn{x} and/or \eqn{y} axes.
}

