% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-draws.R
\name{example-draws}
\alias{example-draws}
\alias{example_group_data}
\alias{example_mcmc_draws}
\alias{example_x_data}
\alias{example_y_data}
\alias{example_yrep_draws}
\title{Example draws to use in demonstrations and tests}
\usage{
example_mcmc_draws(chains = 4, params = 4)

example_yrep_draws()

example_y_data()

example_x_data()

example_group_data()
}
\arguments{
\item{chains}{An interger between 1 and 4 indicating the desired number of
chains.}

\item{params}{An integer between 1 and 6 indicating the desired number of
parameters.}
}
\value{
For \code{example_mcmc_draws}, if \code{chains > 1}, a \code{250}
  (iterations) by \code{chains} by \code{params} array. If \code{chains = 1},
  a \code{250} by \code{params} matrix.

For \code{example_y_data} and \code{example_x_data}, numeric vectors
  of \code{434} observations. For \code{example_group_data}, a factor
  variable with two levels, also containing \code{434} observations.

  For \code{example_yrep_draws}, a \code{1000} (draws) by \code{434} (data
  points) matrix. Each row represents a dataset drawn from the posterior
  predictive distribution of the outcome \code{y} after fitting the linear
  regression model mentioned above.
}
\description{
These functions return example data and MCMC draws that are used in the
examples throught the package documentation. The data and draws are from a
linear regression with outcome \code{y}, intercept \code{alpha}, regression
coefficients \code{beta[1]}, ..., \code{beta[4]}, and residual standard
deviation \code{sigma}.
}
\details{
\describe{
\item{\code{example_mcmc_draws}}{
 MCMC draws from the posterior distribution of the parameters in the linear
 regression model described above. If \code{params = 1} then only the draws
 for the intercept \code{alpha} are included in the returned object. If
 \code{params >= 2}, then draws for \code{sigma}, the residual standard
 deviation, are also included. If \code{params} is between \code{3} and the
 maximum of \code{6}, then draws for regression coefficients \code{beta[k]}
 (\code{k} in \code{1:(params-2)}) are also included.
}
\item{\code{example_y_data}}{
 A numeric vector of observations of the outcome variable in the
 linear regression model.
}
\item{\code{example_x_data}}{
 A numeric vector of observations of one of the predictor variables in the
 linear regression model.
}
\item{\code{example_group_data}}{
 A factor variable with two levels that divides the outcome variable into two
 groups.
}
\item{\code{example_yrep_draws}}{
 Draws from the posterior predictive distribution of the outcome.
}
}
}
\examples{
draws <- example_mcmc_draws()
dim(draws)
dimnames(draws)

draws <- example_mcmc_draws(1, 2)
dim(draws)
colnames(draws)

draws <- example_mcmc_draws(params = 6)
dimnames(draws)[[3]]

y <- example_y_data()
x <- example_x_data()
group <- example_group_data()
length(y)
length(x)
length(group)
tail(data.frame(y, x, group), 5)

yrep <- example_yrep_draws()
dim(yrep) # ncol(yrep) = length(y) = length(x) = length(group)

}
\keyword{internal}

