% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-draws.R
\name{example_mcmc_draws}
\alias{example_mcmc_draws}
\title{Example MCMC draws to use in examples and tests}
\usage{
example_mcmc_draws(chains = 4, params = 4)
}
\arguments{
\item{chains}{An interger between 1 and 4 indicating the desired number of
chains.}

\item{params}{An integer between 1 and 6 indicating the desired number of
parameters.}
}
\value{
A \code{500} (iterations) by \code{chains} by \code{params} array.
}
\description{
Example MCMC draws to use in examples and tests. The draws are from the
posterior distribution of the parameters of a linear regression model.
}
\details{
If \code{params = 1} then only the draws for the intercept
  \code{alpha} are included in the returned object. If \code{params >= 2},
  then draws for \code{sigma}, the residual standard deviation, are also
  included. If \code{params} is between \code{3} and the maximum of \code{6},
  then draws for regression coefficients \code{beta[k]} (\code{k} in
  \code{1:(params-2)}) are also included.
}
\examples{
x <- example_mcmc_draws()
dim(x)
dimnames(x)

x <- example_mcmc_draws(1, 2)
dim(x)
colnames(x)

x <- example_mcmc_draws(params = 6)
dimnames(x)[[3]]

}

