% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{extractors}
\alias{extractors}
\alias{log_posterior}
\alias{log_posterior.stanfit}
\alias{log_posterior.stanreg}
\alias{n_eff}
\alias{n_eff.stanfit}
\alias{n_eff.stanreg}
\alias{nuts_params}
\alias{nuts_params.stanfit}
\alias{nuts_params.stanreg}
\alias{r_hat}
\alias{r_hat.stanfit}
\alias{r_hat.stanreg}
\title{Generics and methods for extracting quantities needed for plotting from model
objects}
\usage{
log_posterior(object, ...)

nuts_params(object, ...)

r_hat(object, ...)

n_eff(object, ...)

\method{log_posterior}{stanfit}(object, inc_warmup = FALSE, ...)

\method{log_posterior}{stanreg}(object, inc_warmup = FALSE, ...)

\method{nuts_params}{stanfit}(object, pars = NULL, inc_warmup = FALSE, ...)

\method{nuts_params}{stanreg}(object, pars = NULL, inc_warmup = FALSE, ...)

\method{r_hat}{stanfit}(object, pars = NULL, ...)

\method{r_hat}{stanreg}(object, pars = NULL, regex_pars = NULL, ...)

\method{n_eff}{stanfit}(object, pars = NULL, ...)

\method{n_eff}{stanreg}(object, pars = NULL, regex_pars = NULL, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Arguments passed to individual methods.}

\item{inc_warmup}{A logical scalar (defaulting to \code{FALSE}) indicating
whether to include warmup draws, if applicable.}

\item{pars}{An optional character vector of parameter names. For
\code{nuts_params} these will be NUTS sampler parameter names rather than
model parameters. If \code{pars} is omitted all parameters are included.}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}. If neither \code{pars} nor \code{regex_pars} is specified
then the default is to use \emph{all} parameters.}
}
\value{
\describe{
\item{\code{log_posterior}}{
\code{log_posterior} methods should return a molten data frame, typically
(but not necessarily) created by \code{\link[reshape2]{melt}}. If the model
represented by \code{object} was fit via MCMC the molten data frame should
have columns \code{"Iteration"} (integer), \code{"Chain"} (integer), and
\code{"Value"} (numeric), in any order. For models fit using other methods,
\code{log_posterior} methods can return a data frame with a single column
\code{"Value"}.
}
\item{\code{nuts_params}}{
\code{nuts_params} methods should return a molten data frame, typically (but
not necessarily) created by \code{\link[reshape2]{melt}}. The molten data
frame should have columns \code{"Parameter"} (factor), \code{"Iteration"}
(integer), \code{"Chain"} (integer), and \code{"Value"} (numeric), in any
order.
}
\item{\code{r_hat}, \code{n_eff}}{
Methods should return (named) vectors.
}
}
}
\description{
Generics and methods for extracting quantities needed for plotting from model
objects
}

