% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_draws.R
\name{fake_draws}
\alias{fake_draws}
\title{Simulate "fake" parameter draws for examples}
\usage{
fake_draws(n = 500, chains = 4, params = 4)
}
\arguments{
\item{n}{Number of draws for each parameter (per chain, if \code{chains >
1}).}

\item{chains}{Number of chains.}

\item{params}{Number of parameters.}
}
\value{
If \code{chains = 1}, an \code{iter} by \code{params} matrix. If
  \code{chains > 1}, an \code{iter} by \code{chains} by \code{params} array.
}
\description{
Simulate fake ---i.e., Monte Carlo, \emph{not} Markov Chain Monte Carlo---
draws to use in examples and tests.
}
\details{
If \code{params = 1} then the only simulated parameter is
\describe{
   \item{}{\code{alpha = rnorm(n, -1, 1)}}
}
If \code{params > 1} then
\describe{
 \item{}{\code{sigma = rlnorm(n, 0, 1/2)}}
}
is also included. And if \code{params > 2} then a parameter
\describe{
 \item{}{\code{beta[k] = rnorm(n, runif(1, -1, 1), 1/2)}}
}
for each \code{k} in \code{1:(params - 2)} is also included.
}
\examples{
x <- fake_draws()
dim(x)
dimnames(x)

x <- fake_draws(100, 1, 2)
dim(x)
colnames(x)

x <- fake_draws(params = 6)
dimnames(x)[[3]]

}

