% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-scatterplots.R
\name{geom_mcmc_hist}
\alias{geom_mcmc_hist}
\alias{geom_mcmc_dens}
\alias{geom_mcmc_scatter}
\alias{geom_mcmc_hex}
\title{Specialized geoms for mcmc plotting with default bayesplot settings}
\usage{
geom_mcmc_hist(
  ...,
  mapping = set_hist_aes(freq),
  np = NULL,
  fill = NULL,
  color = NULL,
  size = NULL,
  binwidth = NULL,
  bins = NULL,
  breaks = NULL,
  na.rm = TRUE,
  freq = TRUE
)

geom_mcmc_dens(
  ...,
  mapping = NULL,
  np = NULL,
  by_chain = FALSE,
  color_chains = FALSE
)

geom_mcmc_scatter(
  ...,
  mapping,
  np = NULL,
  size = 2.5,
  shape = 21,
  alpha = 0.8,
  color = get_color("dh"),
  fill = get_color("d")
)

geom_mcmc_hex(
  ...,
  mapping = ggplot2::aes_(fill = ~scales::rescale(..density..)),
  np = NULL,
  binwidth = NULL
)
}
\arguments{
\item{...}{Passed to the ggplot2 geom function}

\item{fill, color, size, binwidth, breaks, na.rm}{Passed to the ggplot2 geom function}

\item{freq}{Used to set the mapping when mapping is not manually set}
}
\description{
Specialized geoms for mcmc plotting with default bayesplot settings
}
\section{Functions}{
\itemize{
\item \code{geom_mcmc_dens}: Density plot with settings for bayesplot by
default

\item \code{geom_mcmc_scatter}: Scatter plot with settings for bayesplot by
default

\item \code{geom_mcmc_hex}: Hex plot with settings for bayesplot by
default
}}

