% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-geoms.R
\name{geom_mcmc_pairs_histogram}
\alias{geom_mcmc_pairs_histogram}
\title{Specialized geoms for mcmc plotting with default bayesplot settings}
\usage{
geom_mcmc_pairs_histogram(
  ...,
  mapping = set_hist_aes(freq),
  fill = NULL,
  color = NULL,
  size = NULL,
  binwidth = NULL,
  breaks = NULL,
  na.rm = TRUE,
  freq = TRUE
)
}
\arguments{
\item{...}{Passed to the ggplot2 geom function}

\item{fill, color, size, binwidth, breaks, na.rm}{Passed to the ggplot2 geom function}

\item{freq}{Used to set the mapping when mapping is not manually set}
}
\description{
Specialized geoms for mcmc plotting with default bayesplot settings
}
