% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-hist.R
\name{MCMC-hist}
\alias{MCMC-hist}
\alias{mcmc_hist}
\alias{mcmc_hist.array}
\alias{mcmc_hist.data.frame}
\alias{mcmc_hist.matrix}
\title{Histogram of posterior draws}
\usage{
mcmc_hist(x, ...)

\method{mcmc_hist}{array}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, binwidth = NULL, transformations = list(),
  facet_args = list(), ...)

\method{mcmc_hist}{matrix}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, binwidth = NULL, transformations = list(),
  facet_args = list(), ...)

\method{mcmc_hist}{data.frame}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, binwidth = NULL, transformations = list(),
  facet_args = list(), ...)
}
\arguments{
\item{x}{Posterior draws.}

\item{...}{For the generic, arguments passed to the various methods. For the
methods themselves \code{...} is ignored.}

\item{pars}{An optional character vector of parameter names. If not specified
then the first parameter is used.}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}.}

\item{by_chain}{A logical value indicating whether to plot each chain
separately (\code{TRUE}) or merge the chains (\code{FALSE}, the default).}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{transformations}{A named list specifying transformations to apply
to parameters. The name of each list element should be a parameter name and
the content of each list element should be a function (or any item to match
as a function via \code{\link{match.fun}}, e.g. a string naming a
function).}

\item{facet_args}{Arguments (other than \code{facets}) passed to
\code{\link[ggplot2]{facet_wrap}} (if \code{by_chain} is \code{FALSE}) or
\code{\link[ggplot2]{facet_grid}} (if \code{by_chain} is \code{TRUE}) to
control faceting.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Histogram of posterior draws
}
\seealso{
Other MCMC: \code{\link{MCMC-dens}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-pairs}},
  \code{\link{MCMC-trace}}
}

