% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-pairs.R
\name{MCMC-pairs}
\alias{MCMC-pairs}
\alias{mcmc_pairs}
\title{Pairs plot of MCMC draws}
\usage{
mcmc_pairs(x, pars = NULL, regex_pars = NULL, ...)
}
\arguments{
\item{x}{A 3-D array, matrix, or data frame of posterior draws. If \code{x}
is a 3-D array then the dimensions must be \code{[Iteration, Chain,
Parameter]} in that order. If there is only one Markov chain or if all
chains have already been merged, then \code{x} can be a matrix or
data.frame with a column for each parameter.}

\item{pars}{An optional character vector of parameter names. If neither
\code{pars} nor \code{regex_pars} is specified then the default is to use
\emph{all} parameters.}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}. If neither \code{pars} nor \code{regex_pars} is specified
then the default is to use \emph{all} parameters.}

\item{...}{Currently ignored.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Pairs plot of MCMC draws
}
\seealso{
Other MCMC: \code{\link{MCMC-distributions}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-traces}}
}

