% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-recovery.R
\name{MCMC-recovery}
\alias{MCMC-recovery}
\alias{mcmc_recover}
\title{Compare MCMC estimates to true parameter values}
\usage{
mcmc_recover(x, true, ..., prob = 0.9, point_est = c("median", "mean",
  "none"), fatten = 3, size = 1)
}
\arguments{
\item{x}{A 3-D array, matrix, list of matrices, or data frame of MCMC draws.
The \link{MCMC-overview} page provides details on how to specify each these
allowed inputs.}

\item{true}{A numeric vector of true values for each of the parameters in
\code{x}. There should be one value in \code{true} for each parameter
included in \code{x} and the order of the parameters in \code{true} should
be the same as the order of the parameters in \code{x}.}

\item{...}{Currently unused.}

\item{prob}{The probability mass to include in the interval. The
default is \code{0.9} (90\% interval).}

\item{point_est}{The point estimate to show. Either \code{"median"} (the
default), \code{"mean"}, or \code{"none"}.}

\item{fatten, size}{Passed to \code{\link[ggplot2]{geom_pointrange}}.}
}
\description{
See the \strong{Plot Descriptions} section, below, for details.
}
\seealso{
Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-diagnostics}},
  \code{\link{MCMC-distributions}},
  \code{\link{MCMC-intervals}}, \code{\link{MCMC-nuts}},
  \code{\link{MCMC-overview}},
  \code{\link{MCMC-scatterplots}},
  \code{\link{MCMC-traces}}
}

