% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old-mcmc-areas.R
\name{mcmc_areas_old}
\alias{mcmc_areas_old}
\title{Old version of \code{mcmc_areas}.}
\usage{
mcmc_areas_old(x, pars = character(), regex_pars = character(),
  transformations = list(), ..., prob = 0.5, prob_outer = 1,
  point_est = c("median", "mean", "none"), rhat = numeric(), bw = NULL,
  adjust = NULL, kernel = NULL)
}
\arguments{
\item{x}{A 3-D array, matrix, list of matrices, or data frame of MCMC draws.
The \link{MCMC-overview} page provides details on how to specify each these
allowed inputs.}

\item{pars}{An optional character vector of parameter names. If neither
\code{pars} nor \code{regex_pars} is specified then the default is to use
\emph{all} parameters.}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}.}

\item{transformations}{Optionally, transformations to apply to parameters
before plotting. If \code{transformations} is a function or a single string
naming a function then that function will be used to transform all
parameters. To apply transformations to particular parameters, the
\code{transformations} argument can be a named list with length equal to
the number of parameters to be transformed. Currently only univariate
transformations of scalar parameters can be specified (multivariate
transformations will be implemented in a future release). If
\code{transformations} is a list, the name of each list element should be a
parameter name and the content of each list element should be a function
(or any item to match as a function via \code{\link{match.fun}}, e.g. a
string naming a function). If a function is specified by its name as a
string (e.g. \code{"log"}), then it can be used to construct a new
parameter label for the appropriate parameter (e.g. \code{"log(sigma)"}).
If a function itself is specified (e.g. \code{log} or \code{function(x)
log(x)}) then \code{"t"} is used in the new parameter label to indicate
that the parameter is transformed (e.g. \code{"t(sigma)"}).}

\item{...}{Currently unused.}

\item{prob}{The probability mass to include in the inner interval (for
\code{mcmc_intervals}) or in the shaded region (for \code{mcmc_areas}). The
default is \code{0.5} (50\% interval).}

\item{prob_outer}{The probability mass to include in the outer interval. The
default is \code{0.9} for \code{mcmc_intervals} (90\% interval)
and \code{1} for \code{mcmc_areas}.}

\item{point_est}{The point estimate to show. Either \code{"median"} (the
default), \code{"mean"}, or \code{"none"}.}

\item{rhat}{An optional numeric vector of \eqn{\hat{R}}{Rhat} estimates, with
one element per parameter included in \code{x}. If \code{rhat} is provided,
the intervals/areas and point estimates in the resulting plot are colored
based on \eqn{\hat{R}}{Rhat} value. See \code{\link{rhat}} for methods for
extracting \eqn{\hat{R}}{Rhat} estimates.}

\item{bw, adjust, kernel}{For \code{mcmc_areas}, optional arguments passed to
\code{\link[stats]{density}} to override default kernel density estimation
parameters.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Old version of \code{mcmc_areas}.
}
