% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-dens.R
\name{mcmc_dens}
\alias{mcmc_dens}
\alias{mcmc_dens.array}
\alias{mcmc_dens.data.frame}
\alias{mcmc_dens.matrix}
\title{Kernel density plot of posterior draws}
\usage{
mcmc_dens(x, pars = NULL, regex_pars = NULL, by_chain = FALSE,
  transform = NULL, facet_args = list(), ...)

\method{mcmc_dens}{array}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, transform = NULL, facet_args = list(), ...)

\method{mcmc_dens}{matrix}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, transform = NULL, facet_args = list(), ...)

\method{mcmc_dens}{data.frame}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, transform = NULL, facet_args = list(), ...)
}
\arguments{
\item{x}{Posterior draws.}

\item{pars}{An optional character vector of parameter names. If not specified
then the first parameter is used.}

\item{by_chain}{Separate by chain or merge chains?}

\item{facet_args}{Arguments (other than \code{facets}) passed to
\code{\link[ggplot2]{facet_wrap}} (if \code{by_chain} is \code{FALSE}) or
\code{\link[ggplot2]{facet_grid}} (if \code{by_chain} is \code{TRUE}) to
control faceting.}

\item{...}{Currently unused.}

\item{pars}{An optional \link[=grep]{regular expression} to use for parameter
selection. Can be specified instead of \code{pars} or in addition to
\code{pars}.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Kernel density plot of posterior draws
}
\seealso{
Other MCMC: \code{\link{mcmc_hist}},
  \code{\link{mcmc_trace}}
}

