% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_hist.R
\name{mcmc_hist}
\alias{mcmc_hist}
\alias{mcmc_hist.array}
\alias{mcmc_hist.data.frame}
\alias{mcmc_hist.matrix}
\title{Histogram of posterior draws}
\usage{
mcmc_hist(x, pars = NULL, regex_pars = NULL, by_chain = FALSE,
  transform = NULL, binwidth = NULL, ...)

\method{mcmc_hist}{array}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, binwidth = NULL, transform = NULL, ...)

\method{mcmc_hist}{matrix}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, binwidth = NULL, transform = NULL, ...)

\method{mcmc_hist}{data.frame}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, binwidth = NULL, transform = NULL, ...)
}
\arguments{
\item{x}{Posterior draws.}

\item{pars}{Parameter names.}

\item{regex_pars}{Regular expression.}

\item{by_chain}{Separate by chain or merge chains?}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{...}{Passed to \code{\link[ggplot2]{facet_wrap}} or
\code{\link[ggplot2]{facet_grid}} (e.g. \code{scales}).}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Histogram of posterior draws
}
\seealso{
Other MCMC: \code{\link{mcmc_dens}}
}

