% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-pairs.R
\name{mcmc_pairs}
\alias{mcmc_pairs}
\alias{mcmc_pairs.array}
\alias{mcmc_pairs.data.frame}
\alias{mcmc_pairs.matrix}
\title{Pairs plot of MCMC draws}
\usage{
mcmc_pairs(x, pars = NULL, regex_pars = NULL, ...)

\method{mcmc_pairs}{array}(x, pars = NULL, regex_pars = NULL, ...)

\method{mcmc_pairs}{matrix}(x, pars = NULL, regex_pars = NULL, ...)

\method{mcmc_pairs}{data.frame}(x, pars = NULL, regex_pars = NULL, ...)
}
\arguments{
\item{x}{Posterior draws.}

\item{pars}{An optional character vector of parameter names. If not specified
then the first parameter is used.}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}.}

\item{...}{Currently unused.}

\item{size}{An optional value to override \code{\link[ggplot2]{geom_line}}'s
default line size.}

\item{facet_args}{Arguments (other than \code{facets}) passed to
\code{\link[ggplot2]{facet_grid}} to control faceting.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Pairs plot of MCMC draws
}
\seealso{
Other MCMC: \code{\link{MCMC-overview}},
  \code{\link{mcmc_dens}}, \code{\link{mcmc_hist}},
  \code{\link{mcmc_trace}}
}

