% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-trace.R
\name{mcmc_trace}
\alias{mcmc_trace}
\alias{mcmc_trace.array}
\alias{mcmc_trace.data.frame}
\alias{mcmc_trace.matrix}
\title{Traceplot (time series plot) of MCMC draws}
\usage{
mcmc_trace(x, pars = NULL, regex_pars = NULL, size = NULL, n_warmup = 0,
  inc_warmup = n_warmup > 0, window = NULL, facet_args = list(), ...)

\method{mcmc_trace}{array}(x, pars = NULL, regex_pars = NULL, size = NULL,
  n_warmup = 0, inc_warmup = n_warmup > 0, window = NULL,
  facet_args = list(), ...)

\method{mcmc_trace}{matrix}(x, pars = NULL, regex_pars = NULL,
  size = NULL, n_warmup = 0, inc_warmup = n_warmup > 0, window = NULL,
  facet_args = list(), ...)

\method{mcmc_trace}{data.frame}(x, pars = NULL, regex_pars = NULL,
  size = NULL, n_warmup = 0, inc_warmup = n_warmup > 0, window = NULL,
  facet_args = list(), ...)
}
\arguments{
\item{x}{Posterior draws.}

\item{pars}{An optional character vector of parameter names. If not specified
then the first parameter is used.}

\item{size}{An optional value to override \code{\link[ggplot2]{geom_line}}'s
default line size.}

\item{n_warmup}{The number of warmup iterations included in \code{x}.}

\item{inc_warmup}{Include warmup iterations in the plot?}

\item{window}{Integer vector of length two specifying the limits of a range
of iterations to display.}

\item{facet_args}{Arguments (other than \code{facets}) passed to
\code{\link[ggplot2]{facet_wrap}} to control faceting.}

\item{...}{Currently unused.}

\item{pars}{An optional \link[=grep]{regular expression} to use for parameter
selection. Can be specified instead of \code{pars} or in addition to
\code{pars}.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Traceplot (time series plot) of MCMC draws
}
\seealso{
Other MCMC: \code{\link{mcmc_dens}},
  \code{\link{mcmc_hist}}
}

