% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-scatterplots.R
\name{pairs_setfacet}
\alias{pairs_setfacet}
\title{Generate a function to set the faceting parameters and row selection for data
for \code{mcmc_pairs()} geoms}
\usage{
pairs_setfacet(fname1, fvalue1, fname2, fvalue2, select_rows)
}
\arguments{
\item{fname1, fname2}{facet column names}

\item{fvalue1, fvalue2}{facet values}

\item{select_rows}{A logical vector of rows to keep (note, scalar \code{TRUE} will
keep all rows).}
}
\value{
A function that takes in a data.frame and returns a data.frame with
columns \code{fname1} and \code{fname2} set to \code{fvalue1} and \code{fvalue2}.
}
\description{
Generate a function to set the faceting parameters and row selection for data
for \code{mcmc_pairs()} geoms
}
