% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc-residuals.R
\name{PPC-residuals}
\alias{PPC-residuals}
\alias{ppc_resid_binned}
\alias{ppc_resid_hist}
\alias{ppc_resid_scatter}
\alias{ppc_resid_scatter_avg}
\alias{ppc_resid_scatter_avg_vs_x}
\title{PPC residuals (predictive errors)}
\usage{
ppc_resid_hist(y, yrep, ..., binwidth = NULL, freq = TRUE)

ppc_resid_scatter(y, yrep, ..., size = 2.5, alpha = 0.8)

ppc_resid_scatter_avg(y, yrep, ..., size = 2.5, alpha = 0.8)

ppc_resid_scatter_avg_vs_x(y, yrep, x, ..., size = 2.5, alpha = 0.8)

ppc_resid_binned(y, Ey, ..., size = 1, alpha = 0.25)
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See \strong{Details} for additional
instructions.}

\item{...}{Currently unused.}

\item{binwidth}{An optional value used as the \code{binwidth} argument to
\code{\link[ggplot2]{geom_histogram}} to override the default binwidth.}

\item{freq}{For histograms, \code{freq=TRUE} (the default) puts count on the
y-axis. Setting \code{freq=FALSE} puts density on the y-axis. (For many
plots the y-axis text is off by default. To view the count or density
labels on the y-axis see the \code{\link{yaxis_text}} convenience
function.)}

\item{size, alpha}{For scatterplots, arguments passed to
\code{\link[ggplot2]{geom_point}} to control the appearance of the
points. For the binned residual plot, arguments controlling the size of
the outline and opacity of the shaded region indicating the 2-SE bounds.}

\item{x}{A numeric vector the same length as \code{y}.}

\item{Ey}{A matrix of posterior draws of the linear predictor transformed by
the inverse-link function. For logistic regression models, \code{Ey} would
be a matrix of predicted probabilities (with the same dimensions as the
\code{yrep} used by the other PPC plotting functions).}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Various plots of residuals (or more precisely, \emph{predictive errors})
computed from \code{y} and \code{yrep}. See the \strong{Details} and
\strong{Plot Descriptions} sections, below.
}
\details{
All of these functions (aside from the \code{*_scatter_avg} functions)
compute and plot residuals for each row of the matrix \code{yrep} (or for
\code{ppc_resid_binned} the matrix \code{Ey}), so it is usually a good idea
for \code{yrep} and \code{Ey} to contain only a small number of draws (rows).
See \strong{Examples}, below.

For binomial and Bernoulli data the \code{ppc_resid_binned} function
can be used to generate binned residual plots. Bernoulli data can be input
as a vector of 0s and 1s, whereas for binomial data \code{y} should be a
vector of "success" proportions (not a matrix of "success" and "failure"
counts).
}
\section{Plot descriptions}{

\describe{
  \item{\code{ppc_resid_hist}}{
   A separate histogram is plotted for the predictive errors computed from
   \code{y} and each dataset (row) in \code{yrep}. For this plot \code{yrep}
   should have only a small number of rows.
  }
  \item{\code{ppc_resid_scatter}}{
   A separate scatterplot is displayed for \code{y} vs. the predictive errors
   computed from \code{y} and each dataset (row) in \code{yrep}. For this
   plot \code{yrep} should have only a small number of rows.
  }
  \item{\code{ppc_resid_scatter_avg}}{
   A single scatterplot of \code{y} vs. the average of the errors computed
   from \code{y} and each dataset (row) in \code{yrep}. For each individual
   data point \code{y[n]} the average error is the average of the
   errors for \code{y[n]} computed over the the draws from the posterior
   predictive distribution.
  }
  \item{\code{ppc_resid_scatter_avg_vs_x}}{
   Same as ppc_resid_scatter_avg, except the average is plotted on the
   \eqn{y}-axis and a a predictor variable \code{x} is plotted on the
   \eqn{x}-axis.
  }
  \item{\code{ppc_resid_binned}}{
   Intended for use with binomial data. A separate binned residual plot
   (similar to \code{\link[arm]{binnedplot}}) is generated for each dataset
   (row) in \code{Ey}, the posterior draws of the linear predictor
   transformed by the inverse-link function. For this plot \code{Ey} should
   have only a small number of rows.
  }
}
}
\examples{
y <- example_y_data()
yrep <- example_yrep_draws()
ppc_resid_hist(y, yrep[1:3, ])
ppc_resid_scatter(y, yrep[10:14, ])
ppc_resid_scatter_avg(y, yrep)

x <- example_x_data()
ppc_resid_scatter_avg_vs_x(y, yrep, x)

# ppc_resid_binned with binomial model from rstanarm
\dontrun{
library(rstanarm)
example("example_model", package = "rstanarm")
formula(example_model)

# get linear predictor transformed by inverse-link function
Ey <- posterior_linpred(example_model, transform = TRUE)

# get observed proportion of "successes"
y <- example_model$y  # matrix of "success" and "failure" counts
y <- y[, 1] / rowSums(y)  # proportions

ppc_resid_binned(y, Ey[1:6, ])
}

}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
Other PPCs: \code{\link{PPC-distributions}},
  \code{\link{PPC-overview}},
  \code{\link{PPC-scatterplots}},
  \code{\link{PPC-test-statistics}},
  \code{\link{PPC-time-series}}, \code{\link{PPC-vs-x}}
}

