% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc-time-series.R
\name{PPC-time-series}
\alias{PPC-time-series}
\alias{ppc_ts}
\alias{ppc_ts_grouped}
\title{PPC time series}
\usage{
ppc_ts(y, yrep, time, ..., prob = 0.8, alpha = 0.33, size = 1,
  y_style = c("points", "lines"))

ppc_ts_grouped(y, yrep, time, group, facet_args = list(), ..., prob = 0.8,
  alpha = 0.33, size = 1, y_style = c("points", "lines"))
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See \strong{Details} for additional
instructions.}

\item{time}{An optional numeric vector, the same length as \code{y}. If
\code{time} is not provided then it is set to \code{1:length(y)}. For
\code{ppc_ts} the values in \code{time} must be unique. For
\code{ppc_ts_grouped} times can be repeated.}

\item{...}{Currently unused.}

\item{prob}{A value between 0 and 1 indicating the desired probability mass
to include in the \code{yrep} intervals. The default is 0.8.}

\item{alpha, size}{Arguments passed to \code{\link[ggplot2]{geom_ribbon}}. See
Examples.}

\item{y_style}{A character vector of length 1 or 2 indicating how \code{y}
should be plotted. If \code{c("points","lines")}, the default, then
\code{y} is plotted as connected points. Use only \code{"points"} to omit
the connecting lines, and use only \code{"lines"} to show the lines
without the points.}

\item{group}{A grouping variable (a vector or factor) the same length as
\code{y}. Each value in \code{group} is interpreted as the group level
pertaining to the corresponding value of \code{y}.}

\item{facet_args}{An optional list of  arguments (other than \code{facets})
passed to \code{\link[ggplot2]{facet_wrap}} to control faceting.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
PPCs for time series. Medians and central interval estimates of \code{yrep}
by time, with \code{y} overlaid. See the \strong{Plot Descriptions} section,
below.
}
\details{
For Binomial data, the plots will typically be most useful if
  \code{y} and \code{yrep} contain the "success" proportions (not discrete
  "success" or "failure" counts).
}
\section{Plot Descriptions}{

\describe{
  \item{\code{ppc_ts}}{
   \code{100*prob}\% central intervals for \code{yrep} at each time point,
   with a line through the median of \code{yrep} at each time. Values of
   \code{y} are overlaid as points connected by lines, just points, or just
   lines (depending on the value \code{y_style}).
  }
  \item{\code{ppc_ts_grouped}}{
   Same as \code{ppc_ts} but a separate plot is generated for each level of a
   grouping variable.
  }
}
}
\examples{
y <- rnorm(50)
yrep <- matrix(rnorm(5000, 0, 2), ncol = 50)
ppc_ts(y, yrep)
ppc_ts(y, yrep, y_style = "lines")

color_scheme_set("mix-green-blue")
time <- seq(1, 100, by = 2)
ppc_ts(y, yrep, time, alpha = 1, size = .5)
ppc_ts(y, yrep, time, size = .5, y_style = "lines")
ppc_ts(y, yrep, time, alpha = 1, size = .25, y_style = "lines")

color_scheme_set("pink")
time <- rep(1:10, each = 5)
group <- gl(5, 1, length = 50, labels = LETTERS[1:5])
ppc_ts_grouped(y, yrep, time, group) + xaxis_text(FALSE)

ppc_ts_grouped(
 y, yrep, time, group,
 facet_args = list(scales = "fixed"),
 alpha = 1,
 size = 2
) +
 xaxis_text(FALSE) +
 panel_bg(fill = "gray20")

}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
Other PPCs: \code{\link{PPC-distributions}},
  \code{\link{PPC-overview}}, \code{\link{PPC-residuals}},
  \code{\link{PPC-scatterplots}},
  \code{\link{PPC-test-statistics}}, \code{\link{PPC-vs-x}}
}

