% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc-vs-x.R
\name{PPC-vs-x}
\alias{PPC-vs-x}
\alias{ppc_vs_x}
\alias{ppc_vs_x_grouped}
\title{PPC time series}
\usage{
ppc_vs_x(y, yrep, x, ..., prob = 0.8, y_style = c("both", "points",
  "lines"))

ppc_vs_x_grouped(y, yrep, x, group, ..., facet_args = list(), prob = 0.8,
  y_style = c("both", "points", "lines"))
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See \strong{Details} for additional
instructions.}

\item{x}{A numeric vector the same length as \code{y}.}

\item{...}{Currently unused.}

\item{prob}{A value between 0 and 1 indicating the desired probability mass
to include in the \code{yrep} intervals. The default is 0.8.}

\item{y_style}{Should \code{y} be plotted as points connected by lines, only
the points, or only the lines?}

\item{group}{A grouping variable (a vector or factor) the same length as
\code{y}. Each value in \code{group} is interpreted as the group level
pertaining to the corresponding value of \code{y}.}

\item{facet_args}{An optional list of  arguments (other than \code{facets})
passed to \code{\link[ggplot2]{facet_wrap}} to control faceting.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Medians and central interval estimates of \code{yrep} by value of \code{x},
with \code{y} overlaid.
}
\section{Plot Descriptions}{

\describe{
  \item{\code{ppc_vs_x}}{
   \code{100*prob}\% central intervals for \code{yrep} at each \code{x} value,
   with a line through the median of \code{yrep}. Values of
   \code{y} are overlaid as points connected by lines, just points, or just
   lines (depending on the value \code{y_style}).
  }
  \item{\code{ppc_vs_x_grouped}}{
   Same as \code{ppc_vs_x} but a separate plot is generated for each level of a
   grouping variable.
  }
}
}
\examples{
x <- rnorm(200)
y <- rnorm(200, 1 + 2*x, 2)
yrep <- t(replicate(50, matrix(rnorm(200, 1 + 2*x, 3)), simplify = TRUE))
ppc_vs_x(y, yrep, x)
ppc_vs_x(y, yrep, x, prob = 0.5, y_style = "lines") + axis_ticksize(0.25)

set_color_scheme("green")
group <- gl(5, 1, length = 200, labels = LETTERS[1:5])
ppc_vs_x_grouped(y, yrep, x, group)

# don't force all facets to have same x axis scale
ppc_vs_x_grouped(y, yrep, x, group, y_style = "lines",
                 facet_args = list(scales = "free", ncol = 1))

}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other PPCs: \code{\link{PPC-distributions}},
  \code{\link{PPC-overview}}, \code{\link{PPC-residuals}},
  \code{\link{PPC-scatterplots}},
  \code{\link{PPC-test-statistics}},
  \code{\link{PPC-time-series}}
}

