% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_bars.R
\name{ppc_bars}
\alias{ppc_bars}
\alias{ppc_bars_grouped}
\title{Bar plots for ordinal, categorical and multinomial outcomes}
\usage{
ppc_bars(y, yrep, ..., prob = 0.9, width = 0.9, size = 1, fatten = 3,
  freq = TRUE)

ppc_bars_grouped(y, yrep, group, facet_args = list(), ..., prob = 0.9,
  width = 0.9, size = 1, fatten = 3, freq = TRUE)
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See \strong{Details} for additional
instructions.}

\item{...}{Currently unused.}

\item{prob}{A value between 0 and 1 indicating the desired probability mass
to include in the \code{yrep} intervals. The default is 0.9.}

\item{width}{Passed to \code{\link[ggplot2]{geom_bar}} to control the bar
width.}

\item{size, fatten}{Passed to \code{\link[ggplot2]{geom_pointrange}} to
control the appearance of the \code{yrep} points and intervals.}

\item{freq}{If \code{TRUE} (the default) the y-axis will display counts.
Setting \code{freq=FALSE} will put proportions on the y-axis.}

\item{group}{A grouping variable (a vector or factor) the same length as
\code{y}. Each value in \code{group} is interpreted as the group level
pertaining to the corresponding value of \code{y}.}

\item{facet_args}{An optional list of  arguments (other than \code{facets})
passed to \code{\link[ggplot2]{facet_wrap}} to control faceting.}
}
\description{
Bar plots for ordinal, categorical, and multinomial outcomes. See the
\strong{Plot Descriptions} section below.
}
\details{
For \code{ppc_bars}, the observations \code{y} and predictions
  \code{yrep} must be non-negative integers. \pkg{bayesplot} will validate
  that both \code{y} and \code{yrep} contain only non-negative integer
  values, although they need not be integers in the strict sense of \R's
  \code{\link{integer}} type.
}
\section{Plot Descriptions}{

\describe{
\item{\code{ppc_bars}}{
  Bar plot of \code{y} with \code{yrep} medians and uncertainty intervals
  superimposed on the bars.
}
\item{\code{ppc_bars_grouped}}{
  Same as \code{ppc_bars} but a separate plot (facet) is generated for each
  level of a grouping variable.
}
}
}

