% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_counts.R
\name{ppc_counts}
\alias{ppc_counts}
\title{Bar plots for count, multinomial, and ordinal models}
\usage{
ppc_counts(y, yrep, ..., prob = 0.9)
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See \strong{Details} for additional
instructions.}

\item{prob}{The probability mass to include in the uncertainty interval
around the median predicted counts. Defaults to \code{0.9}.}
}
\description{
Bar plots for count, multinomial, and ordinal models
}
\details{
For \code{ppc_counts}, the observations \code{y} and predictons
  \code{yrep} must be non-negative integers. \pkg{bayesplot} will validate
  that both \code{y} and \code{yrep} contain only non-negative integer values
  (although they need not be integers in the strict sense of \R's
  \code{\link{integer}} type).
}
