% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_dist.R
\name{ppc_dist}
\alias{ppc_dist}
\title{Distributions}
\usage{
ppc_dist(y, yrep, overlay = TRUE, ...)
}
\arguments{
\item{y}{A vector of observed data. See Details.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior 
predictive distribution, where \eqn{S} is the size of the posterior sample 
(or subset of the posterior sample used to generate \code{yrep}) and 
\eqn{N} is the number of observations (the length of \code{y}). 
See Details for additional instructions.}

\item{overlay}{Should distributions be plotted as density estimates overlaid
in a single plot (\code{TRUE}, the default) or as separate histograms
(\code{FALSE})? Especially if \code{overlay} is \code{FALSE}, \code{yrep}
should only contain a small number of draws (\code{nrow(yrep)} should be
small).}

\item{...}{Optional arguments to geoms to control features of the plots
(e.g. \code{binwidth} if the plot is a histogram).}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Compare the empirical distribution of the data \eqn{y} to the distributions
of simulated/replicated data \eqn{y^{rep}}{yrep} from the posterior
predictive distribution.
}
\details{
For binomial (not Bernoulli) data, \code{y} and \code{yrep} should 
  contain the "success" proportions (not "success" or "failure" counts). For 
  ordinal data, \code{y} and \code{yrep} should both contain integers.
}
\examples{
y <- rnorm(50)
yrep <- matrix(rnorm(5000), ncol = 50)
ppc_dist(y, yrep)
ppc_dist(y, yrep[1:8, ], overlay = FALSE)

}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
Other PPCs: \code{\link{ppc_resid}},
  \code{\link{ppc_scatter}}, \code{\link{ppc_stat}}
}

