% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_resid.R
\name{ppc_resid}
\alias{ppc_resid}
\alias{ppc_resid_binned}
\title{Residuals}
\usage{
ppc_resid(y, yrep, ...)

ppc_resid_binned(y, Ey, ...)
}
\arguments{
\item{y}{A vector of observed data. See Details.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior 
predictive distribution, where \eqn{S} is the size of the posterior sample 
(or subset of the posterior sample used to generate \code{yrep}) and 
\eqn{N} is the number of observations (the length of \code{y}). 
See Details for additional instructions.}

\item{...}{For \code{ppc_resid}, optional arguments to
\code{\link[ggplot2]{geom_histogram}} (e.g. \code{binwidth}) to control the
appearance of the plot.}

\item{Ey}{A matrix of posterior draws of the linear predictor transformed by
the inverse-link function.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
\code{ppc_resid} plots the distributions of residuals computed from
\eqn{y} and simulated datasets \eqn{y^{rep}}{yrep}. For binomial data,
\code{ppc_resid_binned} generates binned residual plots (similar to
\code{\link[arm]{binnedplot}}) from \eqn{y} and the posterior draws of the
linear predictor transformed by the inverse-link function.
}
\details{
\code{ppc_resid} and \code{ppc_resid_binned} compute and plot
residuals for each row of the matrices \code{yrep} and \code{Ey},
respectively, so it is usually a good idea to \code{yrep} and \code{Ey} to
contain only a small number of draws (rows).

For binomial and Bernoulli data the \code{ppc_resid_binned} function
should be used to generate binned residual plots. Bernoulli data can be input
as a vector of 0s and 1s, whereas for binomial data \code{y} should be a
vector of "success" proportions (not a matrix of "success" and "failure"
counts).
}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
Other PPCs: \code{\link{ppc_dist}},
  \code{\link{ppc_scatter}}, \code{\link{ppc_stat}}
}

