% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_stat.R
\name{ppc_stat}
\alias{ppc_stat}
\title{Test statistics}
\usage{
ppc_stat(y, yrep, stat = "mean", ...)
}
\arguments{
\item{y}{A vector of observed data. See Details.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior 
predictive distribution, where \eqn{S} is the size of the posterior sample 
(or subset of the posterior sample used to generate \code{yrep}) and 
\eqn{N} is the number of observations (the length of \code{y}). 
See Details for additional instructions.}

\item{stat}{A character vector (of length 1 or 2) naming a single function or
a pair of functions. The function(s) should take a vector input and return
a scalar test statistic. If \code{stat} specifies a single function then
the resulting plot is a histogram of \eqn{{T(y^{rep})}}{T(yrep)} and the
value of the test statistic in the observed data, \eqn{T(y)}, is shown in
the plot as a vertical line. If two functions are specified then the plot
is a scatterplot and \eqn{T(y)} is shown as a large point.}

\item{...}{Optional arguments to \code{\link[ggplot2]{geom_histogram}} (if
\code{stat} is a single function) or \code{\link[ggplot2]{geom_point}} (if
\code{stat} specifies two functions).}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
The distribution of a single test statistic \eqn{{T(y^{rep})}}{T(yrep)}, or a
pair of test statistics over the \code{nreps} simulated datasets, compared
to the observed value \eqn{T(y)}.
}
\details{
For binomial (not Bernoulli) data, \code{y} and \code{yrep} should 
  contain the "success" proportions (not "success" or "failure" counts). For 
  ordinal data, \code{y} and \code{yrep} should both contain integers.
}
\examples{
y <- rnorm(30)
yrep <- matrix(rnorm(3000), ncol = 30)
ppc_stat(y, yrep)
ppc_stat(y, yrep, stat = "var")
ppc_stat(y, yrep, stat = c("mean", "sd"))

# define a custom test statistic
q25 <- function(y) quantile(y, 0.25)
ppc_stat(y, yrep, stat = "q25")

}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
Other PPCs: \code{\link{ppc_dist}},
  \code{\link{ppc_resid}}, \code{\link{ppc_scatter}}
}

