% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-group.R
\name{ppcs-by-group}
\alias{ppcs-by-group}
\title{Index of \code{ppc} functions that stratify by group level}
\description{
Index of functions in the \pkg{ppcheck} package that accept a
  \code{group} argument.
}
\section{Functions}{

\describe{
  \item{\code{\link{ppc_scatter_avg_grouped}}}{
   Scatterplots of \code{y} vs \code{yrep} at each level of a grouping
   variable.
  }
  \item{\code{\link{ppc_stat_grouped}}}{
   The posterior predictive distributions of a test statistic
   \eqn{T(y^{rep})}{T(yrep)} compared to the
   observed value \eqn{T(y)} at each level of a grouping variable.
  }
  \item{\code{\link{ppc_violin_grouped}}}{
    Density estimates of \code{yrep} by level of a grouping variable,
    with \code{y} points overlaid.
  }
}
}

