% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppd-distributions.R
\name{PPD-distributions}
\alias{PPD-distributions}
\alias{ppd_data}
\alias{ppd_dens_overlay}
\alias{ppd_ecdf_overlay}
\alias{ppd_dens}
\alias{ppd_hist}
\alias{ppd_dots}
\alias{ppd_freqpoly}
\alias{ppd_freqpoly_grouped}
\alias{ppd_boxplot}
\title{PPD distributions}
\usage{
ppd_data(ypred, group = NULL)

ppd_dens_overlay(
  ypred,
  show_marginal = FALSE,
  ...,
  size = 0.25,
  alpha = 0.7,
  trim = FALSE,
  bw = NULL,
  adjust = NULL,
  kernel = NULL,
  bounds = NULL,
  n_dens = NULL
)

ppd_ecdf_overlay(
  ypred,
  show_marginal = FALSE,
  ...,
  discrete = deprecated(),
  pad = TRUE,
  size = 0.25,
  alpha = 0.7
)

ppd_dens(
  ypred,
  show_marginal = FALSE,
  ...,
  trim = FALSE,
  size = 0.5,
  alpha = 1,
  bounds = NULL
)

ppd_hist(
  ypred,
  show_marginal = FALSE,
  ...,
  binwidth = NULL,
  bins = NULL,
  breaks = NULL,
  freq = !show_marginal
)

ppd_dots(
  ypred,
  show_marginal = FALSE,
  ...,
  binwidth = NA,
  quantiles = 100,
  freq = TRUE
)

ppd_freqpoly(
  ypred,
  show_marginal = FALSE,
  ...,
  binwidth = NULL,
  bins = NULL,
  freq = !show_marginal,
  size = 0.5,
  alpha = 1
)

ppd_freqpoly_grouped(
  ypred,
  group,
  show_marginal = FALSE,
  ...,
  binwidth = NULL,
  bins = NULL,
  freq = !show_marginal,
  size = 0.5,
  alpha = 1
)

ppd_boxplot(
  ypred,
  show_marginal = FALSE,
  ...,
  notch = TRUE,
  size = 0.5,
  alpha = 1
)
}
\arguments{
\item{ypred}{An \code{S} by \code{N} matrix of draws from the posterior (or prior)
predictive distribution, or a \code{\link[posterior:draws]{posterior::draws}} object. The number of
rows, \code{S}, is the size of the posterior (or prior) sample used to generate
\code{ypred}. The number of columns, \code{N}, is the number of predicted
observations.}

\item{group}{A grouping variable of the same length as \code{y}.
Will be coerced to \link[base:factor]{factor} if not already a factor.
Each value in \code{group} is interpreted as the group level pertaining
to the corresponding observation.}

\item{show_marginal}{Plot the marginal PPD along with the \code{ypred}s.}

\item{...}{For dot plots, optional additional arguments to pass to \code{\link[ggdist:stat_dots]{ggdist::stat_dots()}}.}

\item{size, alpha}{Passed to the appropriate geom to control the appearance of
the predictive distributions.}

\item{trim}{A logical scalar passed to \code{\link[ggplot2:geom_density]{ggplot2::geom_density()}}.}

\item{bw, adjust, kernel, n_dens, bounds}{Optional arguments passed to
\code{\link[stats:density]{stats::density()}} (and \code{bounds} to \code{\link[ggplot2:stat_density]{ggplot2::stat_density()}}) to override
default kernel density estimation parameters or truncate the density
support. If \code{NULL} (default), \code{bw} is set to \code{"nrd0"}, \code{adjust} to \code{1},
\code{kernel} to \code{"gaussian"}, and \code{n_dens} to \code{1024}.}

\item{discrete}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} The \code{discrete} argument is
deprecated. The ECDF is a step function by definition, so \code{geom_step()} is
now always used.}

\item{pad}{A logical scalar passed to \code{\link[ggplot2:stat_ecdf]{ggplot2::stat_ecdf()}}.}

\item{binwidth}{Passed to \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}}, \code{\link[ggplot2:geom_area]{ggplot2::geom_area()}},
and \code{\link[ggdist:stat_dots]{ggdist::stat_dots()}} to override the default binwidth.}

\item{bins}{Passed to \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}} and \code{\link[ggplot2:geom_area]{ggplot2::geom_area()}}
to override the default binning.}

\item{breaks}{Passed to \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}} as an alternative to
\code{binwidth}.}

\item{freq}{For histograms and frequency polygons, \code{freq=TRUE} (the default)
puts count on the y-axis. Setting \code{freq=FALSE} puts density on the y-axis.
(For many plots the y-axis text is off by default. To view the count or
density labels on the y-axis see the \code{\link[=yaxis_text]{yaxis_text()}} convenience
function.)}

\item{quantiles}{For dot plots, an optional integer passed to
\code{\link[ggdist:stat_dots]{ggdist::stat_dots()}} specifying the number of quantiles to use for a
quantile dot plot. If \code{quantiles} is \code{NA} then all data points are plotted.
The default is \code{quantiles=100} so that each dot represent one percent of
posterior mass.}

\item{notch}{For the box plot, a logical scalar passed to
\code{\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot()}}. Note: unlike \code{geom_boxplot()}, the default is
\code{notch=TRUE}.}
}
\value{
The plotting functions return a ggplot object that can be further
customized using the \strong{ggplot2} package. The functions with suffix
\verb{_data()} return the data that would have been drawn by the plotting
function.
}
\description{
Plot posterior or prior predictive distributions. Each of these functions
makes the same plot as the corresponding \code{\link[=PPC-distributions]{ppc_}} function
but without plotting any observed data \code{y}. The \strong{Plot Descriptions} section
at \link{PPC-distributions} has details on the individual plots.
}
\details{
For Binomial data, the plots may be more useful if
the input contains the "success" \emph{proportions} (not discrete
"success" or "failure" counts).
}
\examples{
# difference between ppd_dens_overlay() and ppc_dens_overlay()
color_scheme_set("brightblue")
preds <- example_yrep_draws()
ppd_dens_overlay(ypred = preds[1:50, ])
ppd_dens_overlay(ypred = preds[1:50, ], show_marginal = TRUE)
ppc_dens_overlay(y = example_y_data(), yrep = preds[1:50, ])

}
\references{
Säilynoja, T., Johnson, A., Martin, O., and Vehtari, A. (2026).
Recommendations for visual predictive checks in Bayesian workflow.
\emph{Journal of Visualization and Interaction}. 1(1).
doi:10.54337/jovi.v1i1.11478.
}
\seealso{
Other PPDs:
\code{\link{PPD-intervals}},
\code{\link{PPD-overview}},
\code{\link{PPD-test-statistics}}
}
\concept{PPDs}
