% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{Residuals}
\alias{Residuals}
\alias{ppc_resid}
\alias{ppc_resid_binned}
\title{Residuals}
\usage{
ppc_resid(y, yrep, ..., binwidth = NULL)

ppc_resid_binned(y, Ey, ...)
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See \strong{Details} for additional
instructions.}

\item{...}{Currently unused.}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{Ey}{A matrix of posterior draws of the linear predictor transformed by
the inverse-link function.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
\code{ppc_resid} plots the distributions of residuals computed from
\code{y} and simulated datasets \code{yrep}. For binomial data,
\code{ppc_resid_binned} generates binned residual plots (similar to
\code{\link[arm]{binnedplot}}) from \code{y} and the posterior draws of the
linear predictor transformed by the inverse-link function.
}
\details{
\code{ppc_resid} and \code{ppc_resid_binned} compute and plot
residuals for each row of the matrices \code{yrep} and \code{Ey},
respectively, so it is usually a good idea for \code{yrep} and \code{Ey} to
contain only a small number of draws (rows).

For binomial and Bernoulli data the \code{ppc_resid_binned} function
should be used to generate binned residual plots. Bernoulli data can be input
as a vector of 0s and 1s, whereas for binomial data \code{y} should be a
vector of "success" proportions (not a matrix of "success" and "failure"
counts).
}
\section{Plot descriptions}{

\describe{
  \item{\code{ppc_resid}}{
   A separate histogram is plotted for the residuals computed from \code{y}
   and each dataset (row) in \code{yrep}. For this plot
   \code{yrep} should have only a small number of rows.
  }
  \item{\code{ppc_resid_binned}}{
   Intended for use with binomial data. A separate binned residual plot
   (similar to \code{\link[arm]{binnedplot}}) is generated for each dataset
   (row) in \code{Ey}, the posterior draws of the linear predictor
   transformed by the inverse-link function. For this plot \code{Ey} should
   have only a small number of rows.
  }
}
}
\examples{
y <- rnorm(100)
yrep <- matrix(rnorm(2500), ncol = 100)
ppc_resid(y, yrep[1:3, ])
ppc_resid(y, yrep[10:15, ])

}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other PPCs: \code{\link{Distributions}},
  \code{\link{Scatterplots}}, \code{\link{TestStatistics}},
  \code{\link{TimeSeries}}
}

