% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbc.R
\name{sbc_hist}
\alias{sbc_hist}
\title{Diagnostics for Simulation Based Calibration}
\usage{
sbc_hist(ranks, ..., thin = 4, per_bin = 4, worst = 16,
  prob = 0.99, facet_args = list())
}
\arguments{
\item{ranks}{A list of matrices. See \strong{Details}.}

\item{thin}{An integer indicating the thinning interval to use when plotting
so that the histograms consist of (close to) independent realizations. Set
the \code{thin} argument such that the resulting number of draws approximately
matches the effective sample size.}

\item{per_bin}{An integer indicating the number of histogram entries to
combine into a single bar. For best results, one plus the number of draws
from the posterior should be evenly divisible by the number of histogram
bins after thinning. For example, if there are \code{511} posterior draws (per
matrix in \code{ranks}) and \code{thin=4}, then after thinning there will be \code{128}
draws. If \code{per_bin=4}, then \code{128/4=32} histogram bars will be drawn.}

\item{worst}{How many parameters to show, or \code{NA} to plot all parameters. If
\code{worst} is not \code{NA} then parameters are ordered by the degree of
non-uniformity so, for example, \code{worst = 10} means to plot only the \code{10}
worst parameters.}

\item{prob}{The size of the interval plotted to show the expected behavior
under uniformity. The default is \code{prob=0.99}.}

\item{facet_args}{A named list of arguments (other than \code{facets}) passed
to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}} or \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}
to control faceting.}
}
\value{
A ggplot object that can be further customized using the \strong{ggplot2} package.
}
\description{
Diagnostics for Simulation Based Calibration
}
\details{
Each element of the list \code{ranks} should be a matrix of rank comparison
results (encoded as \code{0} or \code{1}) associated with a single draw from the prior
distribution. The columns of each matrix are model parameters and the rows
are posterior draws from the model fit to the data generated from the
corresponding prior draw. An element is \code{0} or \code{1} depending on whether the
posterior draw is greater or less than the corresponding "true" realization
(from the prior). The matrices should have column names to correctly label
the parameters.
}
\section{Plot Descriptions}{

\describe{
\item{\code{sbc_hist()}}{
SBC histogram from Talts et al. (2018). A separate plot (facet) is created
for each parameter.
}
}
}

\examples{
# create some fake inputs to use for sbc_hist()
set.seed(19)
pars <- paste0("beta[", 1:4, "]")
samples_per_prior <- 511
n_replications <- 500
ranks <- list()
for (n in 1:n_replications) {
  r1 <- matrix(0, nrow=samples_per_prior, ncol=length(pars),
               dimnames=list(NULL, pars))
  for (p1 in 1:length(pars)) {
    r1[sample.int(samples_per_prior, floor(runif(1, 0, samples_per_prior))), p1] <- 1
  }
  ranks[[n]] <- r1
}

color_scheme_set("purple")
sbc_hist(ranks)

}
\references{
Talts, S., Betancourt, M., Simpson, D., Vehtari, A., and Gelman, A. (2018).
Validating Bayesian Inference Algorithms with Simulation-Based Calibration.
arXiv preprint arXiv:1804.06788. \url{https://arxiv.org/abs/1804.06788}
}
\seealso{
\code{\link[rstan:sbc]{rstan::sbc()}}
}
