% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.R
\name{scatterplots}
\alias{ppc_scatter_average}
\alias{ppc_scatter_multiple}
\alias{scatterplots}
\title{Scatterplots}
\usage{
ppc_scatter_average(y, yrep, ...)

ppc_scatter_multiple(y, yrep, ...)
}
\arguments{
\item{y}{A vector of observed data. See Details.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior 
predictive distribution, where \eqn{S} is the size of the posterior sample 
(or subset of the posterior sample used to generate \code{yrep}) and 
\eqn{N} is the number of observations (the length of \code{y}). 
See Details for additional instructions.}

\item{...}{Optional arguments to \code{\link[ggplot2]{geom_point}} to control
the appearance of the plotted points.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Scatterplots of the observed data \eqn{y} vs. simulated/replicated data
\eqn{y^{rep}}{yrep} from the posterior predictive distribution. For
\code{ppc_scatter_average}, \eqn{y} is plotted against the average values of
\eqn{y^{rep}}{yrep}, i.e., the points \eqn{(y_n, \bar{y}^{rep}_n),\, n = 1,
\dots, N}{(y_n, mean(yrep_n)), n = 1,...,N}, where each
\eqn{y^{rep}_n}{yrep_n} is a vector of length equal to the number of
posterior draws. For \code{ppc_scatter_multiple}, each of the
\code{nrow(yrep)} simulated datasets is plotted separately against \eqn{y}
(and so \code{yrep} should only contain a small number of draws, i.e.,
\code{nrow(yrep)} should be small).
}
\details{
For binomial (not Bernoulli) data, \code{y} and \code{yrep} should 
  contain the "success" proportions (not "success" or "failure" counts). For 
  ordinal data, \code{y} and \code{yrep} should both contain integers.
}
\examples{
y <- rnorm(100)
yrep <- matrix(rnorm(2500), ncol = 100)
(p1 <- ppc_scatter_average(y, yrep))
(p2 <- ppc_scatter_multiple(y, yrep[1:3, ]))

lims <- ggplot2::lims(x = c(-3, 3), y = c(-3, 3))
p1 + lims
p2 + lims

}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
Other PPCs: \code{\link{distributions}},
  \code{\link{residuals}}, \code{\link{test-statistics}}
}

