% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-themes.R
\name{theme_default}
\alias{theme_default}
\title{bayesplot's default ggplot theme}
\usage{
theme_default(base_size = 11, base_family = "")
}
\arguments{
\item{base_size, base_family}{Passed to \code{\link[ggplot2]{theme_classic}}.}
}
\value{
A ggplot \code{\link[ggplot2]{theme}} object.
}
\description{
A modified version of \code{\link[ggplot2]{theme_classic}}.
}
\seealso{
The \link[=bayesplot-convenience]{bayesplot convenience functions},
  many of which provide shortcuts for tweaking theme elements after creating
  a plot.
}

