% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc-ggplot-theme.R
\name{theme_ppc}
\alias{theme_ppc}
\title{ggplot theme}
\usage{
theme_ppc(x_text = TRUE, y_text = TRUE, x_lab = TRUE, y_lab = TRUE,
  legend_position = "none", ...)
}
\arguments{
\item{x_text, y_text}{Toggle axis text.}

\item{x_lab, y_lab}{Toggle axis labels (titles). Overrided by \code{x_text},
\code{y_text}.}

\item{legend_position}{A valid value to use for \code{legend.position} passed
to \code{\link[ggplot2]{theme}}.}

\item{...}{Other arguments to pass to \code{\link[ggplot2]{theme}}.}
}
\value{
A ggplot \code{\link[ggplot2]{theme}} object.
}
\description{
ggplot theme
}

