% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-series.R
\name{time-series}
\alias{ppc_ts}
\alias{time-series}
\title{PPCs for time series}
\usage{
ppc_ts(y, yrep, time, ..., prob = 0.8, y_style = c("both", "points",
  "lines"))
}
\arguments{
\item{y}{A vector of observed data. See Details.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See Details for additional instructions.}

\item{time}{An optional numeric vector, the same length as \eqn{y}, of
\emph{unique} time values. If \code{time} is missing then it is set to
\code{1:length(y)}.}

\item{...}{Currently unused.}

\item{prob}{A value between 0 and 1 indicating the desired probability mass
to include in the \code{yrep} intervals. The default is 0.8.}

\item{y_style}{Should \code{y} be plotted as points connected by lines, only
the points, or only the lines?}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Central interval estimates (quantiles) of \eqn{y^{rep}}{yrep} by time,
with \eqn{y} overlaid.
}
\examples{
y <- rnorm(50)
yrep <- matrix(rnorm(5000, 0, 2), ncol = 50)
ppc_ts(y, yrep)

time <- sample(200, size = 50)
ppc_ts(y, yrep, time)


}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other PPCs: \code{\link{distributions}},
  \code{\link{residuals}}, \code{\link{scatterplots}},
  \code{\link{test-statistics}}
}

