% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available-module-functions.R
\name{available_ppc}
\alias{available_ppc}
\alias{available_mcmc}
\title{Get or view the names of available plotting functions}
\usage{
available_ppc(pattern)

available_mcmc(pattern)
}
\arguments{
\item{pattern}{An optional \link[=regex]{regular expression}.}
}
\value{
A possibly empty character vector of function names with several
  additional attributes (for use by a custom print method). If \code{pattern}
  is missing then the returned object contains the names of all available
  plotting functions in the \link{MCMC} or \link{PPC} module, depending on
  which function is called. If \code{pattern} is specified then the subset of
  function names matching pattern is returned.
}
\description{
Get or view the names of available plotting functions
}
\examples{
available_mcmc()
available_mcmc("nuts")
available_mcmc("rhat|neff")
available_ppc("grouped")

}
