% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesplot-package.R
\docType{package}
\name{bayesplot-package}
\alias{bayesplot}
\alias{bayesplot-package}
\title{Plotting for Bayesian Models}
\description{
Plotting for Bayesian Models
}
\section{Plotting functionality}{

\itemize{
  \item \strong{\link[=MCMC-overview]{MCMC}}: Visualize Markov chain Monte
  Carlo (MCMC) simulations.
  \item \strong{\link[=PPC-overview]{PPC}}: Perform graphical posterior predictive checking (PPC).
}
}

