% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesplot-package.R
\docType{package}
\name{bayesplot-package}
\alias{bayesplot}
\alias{bayesplot-package}
\title{Graphical posterior predictive checking for Bayesian Models}
\description{
Graphical posterior predictive checking for Bayesian Models
}

