% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesplot-ggplot-themes.R
\name{bayesplot_theme_get}
\alias{bayesplot_theme_get}
\alias{bayesplot_theme_set}
\alias{bayesplot_theme_update}
\alias{bayesplot_theme_replace}
\title{Get, set, and modify the active bayesplot theme}
\usage{
bayesplot_theme_get()

bayesplot_theme_set(new)

bayesplot_theme_update(...)

bayesplot_theme_replace(...)
}
\arguments{
\item{new}{new theme (a list of theme elements)}

\item{...}{named list of theme settings}
}
\description{
Get, set, and modify the active bayesplot theme
}
