% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{distributions}
\alias{distributions}
\alias{ppc_dens}
\alias{ppc_dens_overlay}
\alias{ppc_hist}
\alias{ppc_violin_grouped}
\title{Distributions}
\usage{
ppc_hist(y, yrep, ..., binwidth = NULL)

ppc_dens(y, yrep, ...)

ppc_dens_overlay(y, yrep, ...)

ppc_violin_grouped(y, yrep, group, ...)
}
\arguments{
\item{y}{A vector of observed data. See Details.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See Details for additional instructions.}

\item{...}{Currently unused.}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{group}{A grouping variable of the same length as \code{y}.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Compare the empirical distribution of the data \code{y} to the distributions
of simulated/replicated data \code{yrep} from the posterior predictive
distribution. \code{ppc_dens_overlay} plots the distributions as overlaid
densities. \code{ppc_hist} and \code{ppc_dens} plot separate histograms and
kernel density estimates, respectively, for \code{y} and each row of
\code{yrep} (so for \code{ppc_hist} and \code{ppc_dens} \code{yrep} should
contain only a small number of rows).
}
\details{
For Binomial data, the plots will be most useful if \code{y} and
  \code{yrep} contain the "success" proportions (not "success" or "failure"
  counts).
}
\examples{
y <- rnorm(100)
yrep <- matrix(rnorm(2500), ncol = 100)
ppc_dens_overlay(y, yrep)
ppc_hist(y, yrep[1:8, ])

group <- gl(4, 25, labels = LETTERS[1:4])
ppc_violin_grouped(y, yrep, group)

}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other PPCs: \code{\link{residuals}},
  \code{\link{scatterplots}},
  \code{\link{test-statistics}}, \code{\link{time-series}}
}

