% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-combo.R
\name{MCMC-combos}
\alias{MCMC-combos}
\alias{mcmc_combo}
\title{Combination plots}
\usage{
mcmc_combo(x, ..., combo = c("trace", "dense"), widths = NULL,
  plot = TRUE)
}
\arguments{
\item{x}{A 3-D array, matrix, or data frame of posterior draws. If \code{x}
is a 3-D array then the dimensions must be \code{[Iteration, Chain,
Parameter]} in that order. If there is only one Markov chain or if all
chains have already been merged, then \code{x} can be a matrix or
data.frame with a column for each parameter.}

\item{...}{Arguments passed to the plotting functions named in \code{combo}.}

\item{combo}{A character vector with at least two elements. Each element of
\code{combo} corresponds to a column in the resulting graphic and should be
the name of one of the available \link{MCMC} functions (omitting the
\code{mcmc_} prefix).}

\item{widths}{A numeric vector the same length as \code{combo} specifying
relative column widths. The default, \code{NULL}, allocates the same
horiztonal space for each column.}

\item{plot}{A logical value indicating whether or not to display the plots.
If \code{plot} is \code{FALSE} then the object is returned invisibly
without plotting.}
}
\value{
A gtable object (the result of calling
  \code{\link[gridExtra]{arrangeGrob}}) with \code{length(combo)} columns and
  a row for each parameter.
}
\description{
Combination plots
}
\seealso{
Other MCMC: \code{\link{MCMC-distributions}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-pairs}},
  \code{\link{MCMC-traces}}
}

