% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-diagnostics.R
\name{MCMC-diagnostics}
\alias{MCMC-diagnostics}
\alias{mcmc_neff_dots}
\alias{mcmc_neff_hist}
\alias{mcmc_rhat_dots}
\alias{mcmc_rhat_hist}
\title{Rhat and effective sample size}
\usage{
mcmc_rhat_hist(rhat, ..., binwidth = NULL)

mcmc_rhat_dots(rhat, ..., size = NULL)

mcmc_neff_hist(ratio, ..., binwidth = NULL)

mcmc_neff_dots(ratio, ..., size = NULL)
}
\arguments{
\item{rhat}{Vector of \code{\link[=rhat]{Rhat}} estimates.}

\item{...}{Currently ignored.}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{size}{An optional value to override \code{\link[ggplot2]{geom_point}}'s
default size.}

\item{ratio}{Vector of ratios of effective sample size (estimates) to total
sample size. See \code{\link{neff_ratio}}.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Rhat and effective sample size
}
\section{Plot Descriptions}{

\describe{
  \item{\code{mcmc_rhat_hist, mcmc_rhat_dots}}{
  }
  \item{\code{mcmc_neff_hist, mcmc_neff_dots}}{
  }
  \item{\code{mcmc_diagnostics}}{
  }
}
}
\examples{

# fake rhat values to use for demonstration
rhat <- c(runif(100, 1, 1.15))
mcmc_rhat_hist(rhat)
mcmc_rhat_dots(rhat)

set_color_scheme("blue")
mcmc_rhat_dots(runif(1000, 1, 1.3))
mcmc_rhat_dots(runif(1000, 1, 1.07))


# fake neff ratio values to use for demonstration
ratio <- c(runif(100, 0, 1))
mcmc_neff_hist(ratio)
mcmc_neff_dots(ratio)
mcmc_neff_dots(ratio) + move_legend("top")



}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-distributions}},
  \code{\link{MCMC-intervals}}, \code{\link{MCMC-nuts}},
  \code{\link{MCMC-overview}},
  \code{\link{MCMC-scatterplots}},
  \code{\link{MCMC-traces}}
}

