% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-diagnostics.R
\name{MCMC-diagnostics}
\alias{MCMC-diagnostics}
\alias{mcmc_rhat_dot}
\alias{mcmc_rhat_hist}
\title{Rhat, effective sample size, Monte Carlo standard error}
\usage{
mcmc_rhat_hist(rhat, ..., binwidth = NULL)

mcmc_rhat_dot(rhat, ..., size = NULL)
}
\arguments{
\item{rhat}{Vector of \code{\link[=r_hat]{Rhat}} estimates.}

\item{...}{Currently ignored.}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{size}{An optional value to override \code{\link[ggplot2]{geom_point}}'s
default size.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Rhat, effective sample size, Monte Carlo standard error
}
\section{Plot Descriptions}{

\describe{
  \item{\code{mcmc_rhat}}{
  }
  \item{\code{mcmc_neff}}{
  }
  \item{\code{mcmc_mcse}}{
  }
  \item{\code{mcmc_diagnostics}}{
  }
}
}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-distributions}},
  \code{\link{MCMC-intervals}}, \code{\link{MCMC-nuts}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-pairs}},
  \code{\link{MCMC-traces}}
}

