% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-diagnostics.R
\name{MCMC-diagnostics}
\alias{MCMC-diagnostics}
\alias{mcmc_rhat_dots}
\alias{mcmc_rhat_hist}
\title{Rhat, effective sample size, Monte Carlo standard error}
\usage{
mcmc_rhat_hist(rhat, ..., binwidth = NULL)

mcmc_rhat_dots(rhat, ..., size = NULL)
}
\arguments{
\item{rhat}{Vector of \code{\link[=r_hat]{Rhat}} estimates.}

\item{...}{Currently ignored.}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{size}{An optional value to override \code{\link[ggplot2]{geom_point}}'s
default size.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Rhat, effective sample size, Monte Carlo standard error
}
\section{Plot Descriptions}{

\describe{
  \item{\code{mcmc_rhat_hist, mcmc_rhat_dot}}{
  }
  \item{\code{mcmc_neff_hist, mcmc_neff_dot}}{
  }
  \item{\code{mcmc_mcse_hist, mcmc_mcse_dot}}{
  }
  \item{\code{mcmc_diagnostics}}{
  }
}
}
\examples{

# fake rhat values to use for demonstration
rhat <- c(runif(100, 1, 1.15))
mcmc_rhat_hist(rhat)
mcmc_rhat_dots(rhat)

set_color_scheme("blue")
mcmc_rhat_dots(runif(1000, 1, 1.3))
mcmc_rhat_dots(runif(1000, 1, 1.07)) + ggplot2::ylim(1,1.1)

}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-distributions}},
  \code{\link{MCMC-intervals}}, \code{\link{MCMC-nuts}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-pairs}},
  \code{\link{MCMC-traces}}
}

