% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-diagnostics.R
\name{MCMC-diagnostics}
\alias{MCMC-diagnostics}
\alias{mcmc_neff}
\alias{mcmc_neff_hist}
\alias{mcmc_rhat}
\alias{mcmc_rhat_hist}
\title{Rhat and effective sample size}
\usage{
mcmc_rhat(rhat, ..., size = NULL)

mcmc_rhat_hist(rhat, ..., binwidth = NULL)

mcmc_neff(ratio, ..., size = NULL)

mcmc_neff_hist(ratio, ..., binwidth = NULL)
}
\arguments{
\item{rhat}{A vector of \code{\link[=rhat]{Rhat}} estimates.}

\item{...}{Currently ignored.}

\item{size}{An optional value to override \code{\link[ggplot2]{geom_point}}'s
default size.}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{ratio}{A vector of \emph{ratios} of effective sample size estimates to
total sample size. See \code{\link{neff_ratio}}.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Plots of Rhat and ratio of effective sample size to total sample size.
}
\section{Plot Descriptions}{

\describe{
\item{\code{mcmc_rhat, mcmc_rhat_hist}}{
Rhat values as either points or a histogram. Values are
colored using different shades (lighter is better):
 \itemize{
   \item \emph{light}: below 1.05 (good)
   \item \emph{mid}: between 1.05 and 1.1 (ok)
   \item \emph{dark}: above 1.1 (too high)
 }
}
\item{\code{mcmc_neff, mcmc_neff_hist}}{
Ratios of effective sample size to total sample size as either points or a
histogram. Values are colored using different shades (lighter is better):
 \itemize{
   \item \emph{light}: between 0.5 and 1 (good)
   \item \emph{mid}: between 0.1 and 0.5 (ok)
   \item \emph{dark}: below 0.1 (maybe too low)
 }
}
}
}
\examples{
# fake rhat values to use for demonstration
rhat <- c(runif(100, 1, 1.15))
mcmc_rhat_hist(rhat)
mcmc_rhat(rhat)

set_color_scheme("blue")
mcmc_rhat(runif(1000, 1, 1.3))
mcmc_rhat(runif(1000, 1, 1.07))

# fake neff ratio values to use for demonstration
ratio <- c(runif(100, 0, 1))
mcmc_neff_hist(ratio)
mcmc_neff(ratio)
mcmc_neff(ratio) + move_legend("top") # add legend above plot

\dontrun{
# Example using rstanarm model (requires rstanarm package)
library(rstanarm)

# intentionally use small 'iter' so there are some
# problems with rhat and neff for demonstration
fit <- stan_glm(mpg ~ ., data = mtcars, iter = 50)
rhats <- rhat(fit)
ratios <- neff_ratio(fit)
mcmc_rhat(rhats)
mcmc_neff(ratios)

# there's a small enough number of parameters in the
# model that we can display their names on the y-axis
mcmc_neff(ratios) + yaxis_text()
}

}
\references{
Gelman, A. and Rubin, D. B. (1992). Inference from iterative
simulation using multiple sequences. \emph{Statistical Science}. 7(4),
457--472.

Stan Development Team. (2016). \emph{Stan Modeling Language Users
Guide and Reference Manual.} \url{http://mc-stan.org/documentation/}
}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-distributions}},
  \code{\link{MCMC-intervals}}, \code{\link{MCMC-nuts}},
  \code{\link{MCMC-overview}},
  \code{\link{MCMC-scatterplots}},
  \code{\link{MCMC-traces}}
}

