% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-distributions.R
\name{MCMC-distributions}
\alias{MCMC-distributions}
\alias{mcmc_dens}
\alias{mcmc_dens_overlay}
\alias{mcmc_hist}
\alias{mcmc_hist_by_chain}
\title{Histograms and kernel density plots of posterior draws}
\usage{
mcmc_hist(x, pars = character(), regex_pars = character(),
  transformations = list(), facet_args = list(), binwidth = NULL, ...)

mcmc_dens(x, pars = character(), regex_pars = character(),
  transformations = list(), facet_args = list(), ...)

mcmc_hist_by_chain(x, pars = character(), regex_pars = character(),
  transformations = list(), facet_args = list(), binwidth = NULL, ...)

mcmc_dens_overlay(x, pars = character(), regex_pars = character(),
  transformations = list(), facet_args = list(), ...)
}
\arguments{
\item{x}{A 3-D array, matrix, or data frame of posterior draws. If \code{x}
is a 3-D array then the dimensions must be \code{[Iteration, Chain,
Parameter]} in that order. If there is only one Markov chain or if all
chains have already been merged, then \code{x} can be a matrix or
data.frame with a column for each parameter.}

\item{pars}{An optional character vector of parameter names. If neither
\code{pars} nor \code{regex_pars} is specified then the default is to use
\emph{all} parameters.}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}. If neither \code{pars} nor \code{regex_pars} is specified
then the default is to use \emph{all} parameters.}

\item{transformations}{A named list specifying transformations to apply
to parameters. The name of each list element should be a parameter name and
the content of each list element should be a function (or any item to match
as a function via \code{\link{match.fun}}, e.g. a string naming a
function).}

\item{facet_args}{Arguments (other than \code{facets}) passed to
\code{\link[ggplot2]{facet_wrap}} (if \code{by_chain} is \code{FALSE}) or
\code{\link[ggplot2]{facet_grid}} (if \code{by_chain} is \code{TRUE}) to
control faceting.}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{...}{Currently ignored.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Histograms and kernel density plots of posterior draws
}
\section{Plot Descriptions}{

\describe{
  \item{\code{mcmc_hist}}{
   Histograms of posterior draws with all chains merged.
  }
  \item{\code{mcmc_dens}}{
   Kernel density plots of posterior draws with all chains merged.
  }
  \item{\code{mcmc_hist_by_chain}}{
   Histograms of posterior draws with chains separated via faceting.
  }
  \item{\code{mcmc_dens_overlay}}{
   Kernel density plots of posterior draws with chains separated but
   overlaid on a single plot.
  }
}
}
\seealso{
Other MCMC: \code{\link{MCMC-overview}},
  \code{\link{MCMC-pairs}}, \code{\link{MCMC-traces}}
}

