% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-distributions.R
\name{MCMC-distributions}
\alias{MCMC-distributions}
\alias{mcmc_dens}
\alias{mcmc_dens_overlay}
\alias{mcmc_hist}
\alias{mcmc_hist_by_chain}
\alias{mcmc_violin}
\title{Histograms and kernel density plots of MCMC draws}
\usage{
mcmc_hist(x, pars = character(), regex_pars = character(),
  transformations = list(), facet_args = list(), ..., binwidth = NULL)

mcmc_dens(x, pars = character(), regex_pars = character(),
  transformations = list(), facet_args = list(), ...)

mcmc_hist_by_chain(x, pars = character(), regex_pars = character(),
  transformations = list(), facet_args = list(), ..., binwidth = NULL)

mcmc_dens_overlay(x, pars = character(), regex_pars = character(),
  transformations = list(), facet_args = list(), ...)

mcmc_violin(x, pars = character(), regex_pars = character(),
  transformations = list(), facet_args = list(), ..., probs = c(0.1, 0.5,
  0.9))
}
\arguments{
\item{x}{A 3-D array, matrix, or data frame of posterior draws:
\itemize{
 \item \strong{3-D array}: An \code{\link{array}} with dimensions
 \code{[Iteration, Chain, Parameter]} in that order.
 \item \strong{matrix}: A \code{\link{matrix}} with one column per
 parameter. If \code{x} is a matrix then there should only be a single
 Markov chain or all chains should already be merged (stacked).
 \item \strong{data frame}: There are two types of data frames allowed.
 Either a data frame with one column per parameter (if only a single chain or
 all chains have already been merged), or a data frame with one column per
 parameter plus an additional column \code{"Chain"} that contains the chain
 number (an integer) corresponding to each row in \code{x}.
 \item \strong{list}: A \code{list} of matrices, where each matrix
 corresponds to a Markov chain. All of the matrices should have the same
 number of iterations (rows) and parameters (columns), and parameters should
 have the same names and be in the same order.
}}

\item{pars}{An optional character vector of parameter names. If neither
\code{pars} nor \code{regex_pars} is specified then the default is to use
\emph{all} parameters.}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}. If neither \code{pars} nor \code{regex_pars} is specified
then the default is to use \emph{all} parameters.}

\item{transformations}{An optional named list specifying transformations to
apply to parameters. The name of each list element should be a parameter
name and the content of each list element should be a function (or any item
to match as a function via \code{\link{match.fun}}, e.g. a string naming a
function). If a function in the list of transformations is specified by its
name as a string (e.g. \code{"log"}), then it can be used to construct a
new parameter label for the appropriate parameter (e.g.
\code{"log(sigma)"}). If a function itself is specified (e.g. \code{log} or
\code{function(x) log(x)}) then \code{"t"} is used in the new parameter
label to indicate that the parameter is transformed (e.g.
\code{"t(sigma)"}).}

\item{facet_args}{Arguments (other than \code{facets}) passed to
\code{\link[ggplot2]{facet_wrap}} (if \code{by_chain} is \code{FALSE}) or
\code{\link[ggplot2]{facet_grid}} (if \code{by_chain} is \code{TRUE}) to
control faceting.}

\item{...}{Currently ignored.}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{probs}{A numeric vector passed to \code{\link[ggplot2]{geom_violin}}'s
\code{draw_quantiles} argument to specify at which quantiles to draw
horizontal lines. Set to \code{NULL} to remove the lines.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Histograms and kernel density plots of MCMC draws
}
\section{Plot Descriptions}{

\describe{
  \item{\code{mcmc_hist}}{
   Histograms of posterior draws with all chains merged.
  }
  \item{\code{mcmc_dens}}{
   Kernel density plots of posterior draws with all chains merged.
  }
  \item{\code{mcmc_hist_by_chain}}{
   Histograms of posterior draws with chains separated via faceting.
  }
  \item{\code{mcmc_dens_overlay}}{
   Kernel density plots of posterior draws with chains separated but
   overlaid on a single plot.
  }
  \item{\code{mcmc_violin}}{
   The density estimate of each chain is plotted as a violin with
   horizontal lines at notable quantiles.
  }
}
}
\examples{
# some fake parameter draws to use for demonstration
x <- fake_draws()
dim(x)
dimnames(x)

##################
### Histograms ###
##################

# histograms of all parameters
mcmc_hist(x)

# histograms of some parameters
mcmc_hist(x, pars = c("alpha", "beta[2]"))
mcmc_hist(x, pars = "sigma", regex_pars = "beta")

# interpret facet labels as plotmath expressions
# (e.g. to get greek letters for parameters)
mcmc_hist(
 x,
 pars = c("alpha", "beta[2]"),
 facet_args = list(labeller = ggplot2::label_parsed)
)

# with sigma on log scale
mcmc_hist(x, transformations = list(sigma = "log"))

# can specify transformation as function(x) log(x) or
# as log (without quotes), but then the label is 't(sigma)'
# instead of 'log(sigma)'
mcmc_hist(x, transformations = list(sigma = log))

# separate histograms by chain
set_color_scheme("blue")
mcmc_hist_by_chain(x, pars = "sigma", regex_pars = "beta")

#################
### Densities ###
#################

mcmc_dens(
 x,
 pars = c("sigma", "beta[2]"),
 facet_args = list(nrow = 2)
)

# separate and overlay chains
mcmc_dens_overlay(
 x,
 pars = c("sigma", "beta[2]"),
 facet_args = list(nrow = 2)
)

# parse facet labels, put them on the y axis, and increase their font size
p <- mcmc_dens_overlay(
 x,
 pars = c("sigma", "beta[2]"),
 facet_args =
   list(nrow = 2, labeller = ggplot2::label_parsed, switch = "y")
)
p + facet_fontsize(15)

# separate chains as violin plots
mcmc_violin(x)
mcmc_violin(
 x,
 probs = c(0.1, 0.9), # where to draw quantile lines
 transformations = list(sigma = "log")
)


}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-diagnostics}},
  \code{\link{MCMC-intervals}}, \code{\link{MCMC-nuts}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-pairs}},
  \code{\link{MCMC-traces}}
}

