% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-intervals.R
\name{MCMC-intervals}
\alias{MCMC-intervals}
\alias{mcmc_areas}
\alias{mcmc_intervals}
\title{Plot interval estimates from MCMC draws}
\usage{
mcmc_intervals(x, pars = character(), regex_pars = character(),
  transformations = list(), ..., prob = 0.5, prob_outer = 0.9,
  point_est = c("median", "mean"), rhat = numeric())

mcmc_areas(x, pars = character(), regex_pars = character(),
  transformations = list(), ..., prob = 0.5, point_est = c("median",
  "mean"), rhat = numeric())
}
\arguments{
\item{x}{A 3-D array, matrix, or data frame of posterior draws. If \code{x}
is a 3-D array then the dimensions must be \code{[Iteration, Chain,
Parameter]} in that order. If there is only one Markov chain or if all
chains have already been merged, then \code{x} can be a matrix or
data.frame with a column for each parameter.}

\item{pars}{An optional character vector of parameter names. If neither
\code{pars} nor \code{regex_pars} is specified then the default is to use
\emph{all} parameters.}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}. If neither \code{pars} nor \code{regex_pars} is specified
then the default is to use \emph{all} parameters.}

\item{transformations}{A named list specifying transformations to apply
to parameters. The name of each list element should be a parameter name and
the content of each list element should be a function (or any item to match
as a function via \code{\link{match.fun}}, e.g. a string naming a
function).}

\item{...}{Currently unused.}

\item{prob}{The probability mass to include in inner interval (for
\code{mcmc_intervals}) or in the shaded region (for \code{mcmc_areas}). The
default is \code{0.5} (50\% interval).}

\item{prob_outer}{For \code{mcmc_intervals} only, the probability mass to
include in outer interval. The default is \code{0.9} for
\code{mcmc_intervals} (90\% interval).}

\item{point_est}{The point estimate to show. Either \code{"median"} (the
default) or \code{"mean"}.}

\item{rhat}{An optional numeric vector of \eqn{\hat{R}}{Rhat} estimates, with
one element per parameter included in \code{x}. If \code{rhat} is provided,
the intervals/areas and point estimates in the resulting plot are colored
based on \eqn{\hat{R}}{Rhat} value.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Plot interval estimates from MCMC draws
}
\section{Plot Descriptions}{

\describe{
  \item{\code{mcmc_intervals}}{
   Plots of uncertainty intervals computed from posterior draws with all
   chains merged.
  }
  \item{\code{mcmc_areas}}{
   Density plots of computed from posterior draws with all chains merged,
   with uncertainty intervals shown as shaded areas under the curves.
  }
}
}
\seealso{
Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-distributions}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-pairs}},
  \code{\link{MCMC-traces}}
}

