% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_intervals.R
\name{MCMC-intervals}
\alias{MCMC-intervals}
\alias{mcmc_intervals}
\title{Plot interval estimates from MCMC draws}
\usage{
mcmc_intervals(x, pars = character(), regex_pars = character(),
  transformations = list(), ..., show_density = FALSE, prob_inner = 0.5,
  prob_outer = 0.95, point_est = c("median", "mean"))
}
\arguments{
\item{x}{A 3-D array, matrix, or data frame of posterior draws. If \code{x}
is a 3-D array then the dimensions must be \code{[Iteration, Chain,
Parameter]} in that order. If there is only one Markov chain or if all
chains have already been merged, then \code{x} can be a matrix or
data.frame with a column for each parameter.}

\item{pars}{An optional character vector of parameter names. If neither
\code{pars} nor \code{regex_pars} is specified then the default is to use
\emph{all} parameters.}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}. If neither \code{pars} nor \code{regex_pars} is specified
then the default is to use \emph{all} parameters.}

\item{transformations}{A named list specifying transformations to apply
to parameters. The name of each list element should be a parameter name and
the content of each list element should be a function (or any item to match
as a function via \code{\link{match.fun}}, e.g. a string naming a
function).}

\item{prob_inner}{Probability mass to include in inner interval.}

\item{prob_outer}{Probability mass to include in outer interval.}

\item{point_est}{The point estimate to show. Either \code{"median"} (the
default) or \code{"mean"}.}
}
\description{
Plot interval estimates from MCMC draws
}
\seealso{
Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-distributions}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-pairs}},
  \code{\link{MCMC-traces}}
}

