% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-alg-nuts.R
\name{MCMC-nuts}
\alias{MCMC-nuts}
\alias{NUTS}
\alias{mcmc_nuts_accept_stat}
\alias{mcmc_nuts_divergent}
\alias{mcmc_nuts_stepsize}
\alias{mcmc_nuts_treedepth}
\title{Plots for the No-U-Turn-Sampler (NUTS)}
\usage{
mcmc_nuts_accept_stat(x, lp, chain = NULL, ..., binwidth = NULL)

mcmc_nuts_treedepth(x, lp, chain = NULL, ...)

mcmc_nuts_divergent(x, lp, chain = NULL, ...)

mcmc_nuts_stepsize(x, lp, chain = NULL, ...)
}
\arguments{
\item{x}{A molten data frame of NUTS sampler parameters, either created by
\code{\link{nuts_params}} or in the same form as the object returned by
\code{\link{nuts_params}}.}

\item{lp}{A molten data frame of log-posterior draws, either created by
\code{\link{log_posterior}} or in the same form as the object returned
by \code{\link{log_posterior}}.}

\item{chain}{A positive integer for selecting a particular chain. The default
(\code{NULL}) is to merge the chains before plotting. If \code{chain = k}
then the plot for chain \code{k} is overlaid on top of the plot for all
chains.}

\item{...}{Currently ignored.}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}
}
\value{
A gtable object (the result of calling
  \code{\link[gridExtra]{arrangeGrob}}) created from several ggplot objects.
}
\description{
Plots for the No-U-Turn-Sampler (NUTS)
}
\examples{
\dontrun{
# Assuming 'fit' is a stanfit (rstan) or stanreg (rstanarm) object
np <- nuts_params(fit)
lp <- log_posterior(fit)

mcmc_nuts_accept_stat(np, lp)
mcmc_nuts_accept_stat(np, lp, chain = 3)
}

}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-diagnostics}},
  \code{\link{MCMC-distributions}},
  \code{\link{MCMC-intervals}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-pairs}},
  \code{\link{MCMC-traces}}
}

