% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-alg-nuts.R
\name{MCMC-nuts}
\alias{MCMC-nuts}
\alias{NUTS}
\alias{mcmc_nuts_accept_stat}
\alias{mcmc_nuts_divergent}
\alias{mcmc_nuts_stepsize}
\alias{mcmc_nuts_treedepth}
\title{Plots for the No-U-Turn-Sampler (NUTS)}
\usage{
mcmc_nuts_accept_stat(accept_stat, lp, chain = NULL, ..., binwidth = NULL)

mcmc_nuts_treedepth(treedepth, accept_stat, lp, chain = NULL, ...)

mcmc_nuts_divergent(divergent, accept_stat, lp, chain = NULL, ...)

mcmc_nuts_stepsize(stepsize, accept_stat, lp, chain = NULL, ...)
}
\arguments{
\item{accept_stat}{NUTS \code{accept_stat} parameter. A list with
one component (a vector) per chain or a matrix with one column per chain.}

\item{lp}{Log-posterior. A list with one component (a vector) per chain or
a matrix with one column per chain.}

\item{chain}{A positive integer for selecting a particular chain. The default
(\code{NULL}) is to merge the chains before plotting. If \code{chain = k}
then the plot for chain \code{k} is overlaid on top of the plot for all
chains.}

\item{...}{Currently ignored.}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{treedepth}{NUTS \code{treedepth} parameter. A list with
one component (a vector) per chain or a matrix with one column per chain.}

\item{divergent}{NUTS \code{divergent} parameter. A list with
one component (a vector) per chain or a matrix with one column per chain.}

\item{stepsize}{NUTS \code{stepsize} parameter. A list with
one component (a vector) per chain or a matrix with one column per chain.}
}
\value{
A gtable object (the result of calling
  \code{\link[gridExtra]{arrangeGrob}}) created from several ggplot objects.
}
\description{
Plots for the No-U-Turn-Sampler (NUTS)
}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-distributions}},
  \code{\link{MCMC-intervals}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-pairs}},
  \code{\link{MCMC-traces}}
}

