% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-alg-nuts.R
\name{MCMC-nuts}
\alias{MCMC-nuts}
\alias{NUTS}
\alias{mcmc_nuts_acceptance}
\alias{mcmc_nuts_divergence}
\alias{mcmc_nuts_energy}
\alias{mcmc_nuts_stepsize}
\alias{mcmc_nuts_treedepth}
\title{Diagnostic plots for the No-U-Turn-Sampler (NUTS)}
\usage{
mcmc_nuts_acceptance(x, lp, chain = NULL, ..., binwidth = NULL)

mcmc_nuts_divergence(x, lp, chain = NULL, ...)

mcmc_nuts_stepsize(x, lp, chain = NULL, ...)

mcmc_nuts_treedepth(x, lp, chain = NULL, ...)

mcmc_nuts_energy(x, ..., binwidth = NULL, alpha = 0.5,
  merge_chains = TRUE)
}
\arguments{
\item{x}{A molten data frame of NUTS sampler parameters, either created by
\code{\link{nuts_params}} or in the same form as the object returned by
\code{\link{nuts_params}}.}

\item{lp}{A molten data frame of log-posterior draws, either created by
\code{\link{log_posterior}} or in the same form as the object returned
by \code{\link{log_posterior}}.}

\item{chain}{A positive integer for selecting a particular chain. The default
(\code{NULL}) is to merge the chains before plotting. If \code{chain = k}
then the plot for chain \code{k} is overlaid (in a darker shade but with
transparency) on top of the plot for all chains. The \code{chain} argument
is not used by \code{mcmc_nuts_energy}.}

\item{...}{Currently ignored.}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{alpha}{For \code{mcmc_nuts_energy} only, the transparency (alpha) level
in [0,1] used for the overlaid histogram.}

\item{merge_chains}{For \code{mcmc_nuts_energy} only, should all chains be
merged or displayed separately?}
}
\value{
A gtable object (the result of calling
  \code{\link[gridExtra]{arrangeGrob}}) created from several ggplot objects,
  except for \code{mcmc_nuts_energy}, which returns a ggplot object.
}
\description{
Diagnostic plots for the No-U-Turn-Sampler (NUTS)
}
\section{Plot Descriptions}{

\describe{
  \item{\code{mcmc_nuts_acceptance}}{
  Three plots:
  \itemize{
   \item Histogram of \code{accept_stat__}, which for NUTS is the the
   average acceptance probabilities of all possible samples in the proposed
   tree (NUTS uses a slice sampling algorithm for rejection).
   \item Histogram of the log-posterior (up to a constant).
   \item Scatterplot of \code{accept_stat__} vs log-posterior.
   }
  }
  \item{\code{mcmc_nuts_divergence}}{
  Two plots:
  \itemize{
   \item Violin plots of \code{log-posterior|divergent__=1} and
     \code{log-posterior|divergent__=0}.
   \item Violin plots of \code{accept_stat__|divergent__=1} and
     \code{log-posterior|divergent__=0}.
   }
  }
  \item{\code{mcmc_nuts_stepsize}}{
  Two plots:
  \itemize{
   \item Violin plots of \code{log-posterior} by chain ordered by
   \code{stepsize__} value.
   \item Violin plots of \code{accept_stat__} by chain ordered by
   \code{stepsize__} value.
   }
  }
  \item{\code{mcmc_nuts_treedepth}}{
  Three plots:
  \itemize{
   \item Violin plots of \code{log-posterior} by value of \code{treedepth__}.
   \item Violin plots of \code{accept_stat__} by value of \code{treedepth__}.
   \item Histogram of \code{treedepth__}.
   }
  }
  \item{\code{mcmc_nuts_energy}}{
  Overlaid histograms showing \code{energy__} vs the change in
  \code{energy__}. See Betancourt (2016) for details.
  }
}
}
\examples{
\dontrun{
library(ggplot2)
library(rstanarm)
fit <- stan_glm(mpg ~ wt + am, data = mtcars, iter = 1000)
np <- nuts_params(fit)
lp <- log_posterior(fit)
mcmc_nuts_acceptance(np, lp)
mcmc_nuts_acceptance(np, lp, chain = 2)

set_color_scheme("blue")
mcmc_nuts_energy(np)
mcmc_nuts_energy(np, binwidth = .25, alpha = .8)
(energy_plot <- mcmc_nuts_energy(np, merge_chains = FALSE))
energy_plot +
 facet_wrap(~ Chain, nrow = 1) +
 coord_fixed(ratio = 150)
}

}
\references{
Betancourt, M. (2016). Diagnosing suboptimal cotangent
  disintegrations in Hamiltonian Monte Carlo.
  \url{https://arxiv.org/abs/1604.00695}

Betancourt, M. and Girolami, M. (2013). Hamiltonian Monte Carlo for
hierarchical models. \url{https://arxiv.org/abs/1312.0906}

Hoffman, M. D. and Gelman, A. (2014). The No-U-Turn Sampler:
  adaptively setting path lengths in Hamiltonian Monte Carlo. \emph{Journal of
  Machine Learning Research}. 15:1593--1623.

Stan Development Team. (2016). \emph{Stan Modeling Language Users
Guide and Reference Manual.} \url{http://mc-stan.org/documentation/}
}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-diagnostics}},
  \code{\link{MCMC-distributions}},
  \code{\link{MCMC-intervals}},
  \code{\link{MCMC-overview}},
  \code{\link{MCMC-scatterplots}},
  \code{\link{MCMC-traces}}
}

