% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-alg-nuts.R
\name{MCMC-nuts}
\alias{MCMC-nuts}
\alias{NUTS}
\alias{mcmc_nuts_acceptance}
\alias{mcmc_nuts_divergence}
\alias{mcmc_nuts_energy}
\alias{mcmc_nuts_stepsize}
\alias{mcmc_nuts_treedepth}
\title{Plots for the No-U-Turn-Sampler (NUTS)}
\usage{
mcmc_nuts_acceptance(x, lp, chain = NULL, ..., binwidth = NULL)

mcmc_nuts_divergence(x, lp, chain = NULL, ...)

mcmc_nuts_stepsize(x, lp, chain = NULL, ...)

mcmc_nuts_treedepth(x, lp, chain = NULL, ...)

mcmc_nuts_energy(x, ..., binwidth = NULL, alpha = 0.5,
  merge_chains = TRUE)
}
\arguments{
\item{x}{A molten data frame of NUTS sampler parameters, either created by
\code{\link{nuts_params}} or in the same form as the object returned by
\code{\link{nuts_params}}.}

\item{lp}{A molten data frame of log-posterior draws, either created by
\code{\link{log_posterior}} or in the same form as the object returned
by \code{\link{log_posterior}}.}

\item{chain}{A positive integer for selecting a particular chain. The default
(\code{NULL}) is to merge the chains before plotting. If \code{chain = k}
then the plot for chain \code{k} is overlaid (in a darker shade but with
transparency) on top of the plot for all chains. The \code{chain} argument
is not used by \code{mcmc_nuts_energy}.}

\item{...}{Currently ignored.}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{alpha}{For \code{mcmc_nuts_energy} only, the transparency (alpha) level
in [0,1] used for the overlaid histogram.}

\item{merge_chains}{For \code{mcmc_nuts_energy} only, should all chains be
merged or displayed separately?}
}
\value{
A gtable object (the result of calling
  \code{\link[gridExtra]{arrangeGrob}}) created from several ggplot objects.
}
\description{
Plots for the No-U-Turn-Sampler (NUTS)
}
\examples{
\dontrun{
library(ggplot2)
library(rstanarm)
fit <- stan_glm(mpg ~ wt + am, data = mtcars, iter = 1000)
np <- nuts_params(fit)
lp <- log_posterior(fit)
mcmc_nuts_acceptance(np, lp)
mcmc_nuts_acceptance(np, lp, chain = 2)

set_color_scheme("blue")
mcmc_nuts_energy(np)
mcmc_nuts_energy(np, binwidth = .25, alpha = .8)
(energy_plot <- mcmc_nuts_energy(np, merge_chains = FALSE))
energy_plot +
 facet_wrap(~ Chain, nrow = 1) +
 coord_fixed(ratio = 150)
}

}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-diagnostics}},
  \code{\link{MCMC-distributions}},
  \code{\link{MCMC-intervals}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-pairs}},
  \code{\link{MCMC-traces}}
}

