% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{MCMC-overview}
\alias{MCMC}
\alias{MCMC-overview}
\title{Plots for Markov chain Monte Carlo simulations}
\description{
The \pkg{bayesplot} package provides various plotting functions for creating
graphical displays of Markov chain Monte Carlo (MCMC) simulations. Future
versions of \pkg{bayesplot} will provide many more MCMC plotting functions.
}
\section{MCMC plotting functions}{


\describe{
  \item{\strong{\link[=MCMC-distributions]{Posterior distributions}}}{
    Histograms and kernel density plots of parameter draws, optionally
    showing each Markov chain separately.
  }
  \item{\strong{\link[=MCMC-traces]{Traceplots}}}{
    Times series of parameter draws.
  }
  \item{\strong{\link[=MCMC-pairs]{Pairs}}}{
    Scatterplot matrix of parameter draws.
  }
}
}
\seealso{
Other MCMC: \code{\link{MCMC-distributions}},
  \code{\link{MCMC-pairs}}, \code{\link{MCMC-traces}}
}

