% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-pairs.R
\name{MCMC-pairs}
\alias{MCMC-pairs}
\alias{mcmc_pairs}
\title{Pairs plot of MCMC draws}
\usage{
mcmc_pairs(x, pars = character(), regex_pars = character(),
  transformations = list(), ...)
}
\arguments{
\item{x}{A 3-D array, matrix, or data frame of posterior draws:
\itemize{
 \item \strong{3-D array}: An \code{\link{array}} with dimensions
 \code{[Iteration, Chain, Parameter]} in that order.
 \item \strong{matrix}: A \code{\link{matrix}} with one column per
 parameter. If \code{x} is a matrix then there should only be a single
 Markov chain or all chains should already be merged (stacked).
 \item \strong{data frame}: There are two types of data frames allowed.
 Either a data frame with one column per parameter (if only a single chain or
 all chains have already been merged), or a data frame with one column per
 parameter plus an additional column \code{"Chain"} that contains the chain
 number (an integer) corresponding to each row in \code{x}.
 \item \strong{list}: A \code{list} of matrices, where each matrix
 corresponds to a Markov chain. All of the matrices should have the same
 number of iterations (rows) and parameters (columns), and parameters should
 have the same names and be in the same order.
}}

\item{pars}{An optional character vector of parameter names. If neither
\code{pars} nor \code{regex_pars} is specified then the default is to use
\emph{all} parameters.}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}. If neither \code{pars} nor \code{regex_pars} is specified
then the default is to use \emph{all} parameters.}

\item{transformations}{An optional named list specifying transformations to
apply to parameters. The name of each list element should be a parameter
name and the content of each list element should be a function (or any item
to match as a function via \code{\link{match.fun}}, e.g. a string naming a
function). If a function in the list of transformations is specified by its
name as a string (e.g. \code{"log"}), then it can be used to construct a
new parameter label for the appropriate parameter (e.g.
\code{"log(sigma)"}). If a function itself is specified (e.g. \code{log} or
\code{function(x) log(x)}) then \code{"t"} is used in the new parameter
label to indicate that the parameter is transformed (e.g.
\code{"t(sigma)"}).}

\item{...}{Currently ignored.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Pairs plot of MCMC draws
}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-diagnostics}},
  \code{\link{MCMC-distributions}},
  \code{\link{MCMC-intervals}}, \code{\link{MCMC-nuts}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-traces}}
}

