% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-parcoord.R
\name{MCMC-parcoord}
\alias{MCMC-parcoord}
\alias{mcmc_parcoord}
\alias{parcoord_style_np}
\title{Parallel coordinates plot of MCMC draws}
\usage{
mcmc_parcoord(x, pars = character(), regex_pars = character(),
  transformations = list(), ..., size = 0.2, alpha = 0.2, np = NULL,
  np_style = parcoord_style_np())

parcoord_style_np(div_color = "red", div_size = 0.2, div_alpha = 0.2)
}
\arguments{
\item{x}{A 3-D array, matrix, list of matrices, or data frame of MCMC draws.
The \link{MCMC-overview} page provides details on how to specify each these
allowed inputs.}

\item{pars}{An optional character vector of parameter names. If neither
\code{pars} nor \code{regex_pars} is specified then the default is to use
\emph{all} parameters.}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}.}

\item{transformations}{Optionally, transformations to apply to parameters
before plotting. If \code{transformations} is a function or a single string
naming a function then that function will be used to transform all
parameters. To apply transformations to particular parameters, the
\code{transformations} argument can be a named list with length equal to
the number of parameters to be transformed. Currently only univariate
transformations of scalar parameters can be specified (multivariate
transformations will be implemented in a future release). If
\code{transformations} is a list, the name of each list element should be a
parameter name and the content of each list element should be a function
(or any item to match as a function via \code{\link{match.fun}}, e.g. a
string naming a function). If a function is specified by its name as a
string (e.g. \code{"log"}), then it can be used to construct a new
parameter label for the appropriate parameter (e.g. \code{"log(sigma)"}).
If a function itself is specified (e.g. \code{log} or \code{function(x)
log(x)}) then \code{"t"} is used in the new parameter label to indicate
that the parameter is transformed (e.g. \code{"t(sigma)"}).}

\item{...}{Currently ignored.}

\item{size, alpha}{Arguments passed on to \code{\link[ggplot2]{geom_line}}.}

\item{np}{For models fit using \code{\link{NUTS}} (more generally,
any \href{http://en.wikipedia.org/wiki/Symplectic_integrator}{symplectic
integrator}), an optional data frame providing NUTS
diagnostic information. The data frame should be the object returned by
\code{\link{nuts_params}} or one with the same structure.}

\item{np_style}{A call to the \code{parcoord_style_np} helper function to
specify arguments controlling the appearance of superimposed lines
representing NUTS diagnostics (in this case divergences) if the \code{np}
argument is specified.}

\item{div_color, div_size, div_alpha}{Optional arguments to the
\code{parcoord_style_np} helper function that are eventually passed to
\code{\link[ggplot2]{geom_line}} if the \code{np} argument is also
specified. They control the color, size, and transparency specifications
for showing divergences in the plot. The default values are displayed in
the \strong{Usage} section above.}
}
\description{
Parallel coordinates plot of MCMC draws. See the \strong{Plot Descriptions}
section, below, for details.
}
\examples{
color_scheme_set("pink")
x <- example_mcmc_draws()
mcmc_parcoord(x)

\dontrun{
library(rstan)
fit <- stan_demo("eight_schools")
draws <- as.array(fit, pars = c("tau", "theta", "lp__"))
np <- nuts_params(fit)

color_scheme_set("blue")
mcmc_parcoord(draws)
mcmc_parcoord(draws, np = np)

color_scheme_set("darkgray")
div_style <- parcoord_style_np(div_color = "green", div_size = 0.5, div_alpha = 0.1)
mcmc_parcoord(draws, size = 0.25, alpha = 0.1,
              np = np, np_style = div_style)
}
}
\seealso{
Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-diagnostics}},
  \code{\link{MCMC-distributions}},
  \code{\link{MCMC-intervals}}, \code{\link{MCMC-nuts}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-recover}},
  \code{\link{MCMC-scatterplots}},
  \code{\link{MCMC-traces}}
}
