% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-scatterplots.R
\name{MCMC-scatterplots}
\alias{MCMC-scatterplots}
\alias{mcmc_pairs}
\alias{mcmc_scatter}
\title{Scatterplots of MCMC draws}
\usage{
mcmc_scatter(x, pars = character(), regex_pars = character(),
  transformations = list(), ..., size = 2.5, alpha = 0.75)

mcmc_pairs(x, pars = character(), regex_pars = character(),
  transformations = list(), size = 1.5, alpha = 0.5, ...)
}
\arguments{
\item{x}{A 3-D array, matrix, or data frame of posterior draws:
\itemize{
 \item \strong{3-D array}: An \code{\link{array}} with dimensions
 \code{[Iteration, Chain, Parameter]} in that order.
 \item \strong{matrix}: A \code{\link{matrix}} with one column per
 parameter. If \code{x} is a matrix then there should only be a single
 Markov chain or all chains should already be merged (stacked).
 \item \strong{data frame}: There are two types of data frames allowed.
 Either a data frame with one column per parameter (if only a single chain or
 all chains have already been merged), or a data frame with one column per
 parameter plus an additional column \code{"Chain"} that contains the chain
 number (an integer) corresponding to each row in \code{x}.
 \item \strong{list}: A \code{list} of matrices, where each matrix
 corresponds to a Markov chain. All of the matrices should have the same
 number of iterations (rows) and parameters (columns), and parameters should
 have the same names and be in the same order.
}}

\item{pars}{An optional character vector of parameter names. (\strong{Note}:
for \code{mcmc_scatter} only two parameters can be selected.)}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}.}

\item{transformations}{An optional named list specifying transformations to
apply to parameters. The name of each list element should be a parameter
name and the content of each list element should be a function (or any item
to match as a function via \code{\link{match.fun}}, e.g. a string naming a
function). If a function in the list of transformations is specified by its
name as a string (e.g. \code{"log"}), then it can be used to construct a
new parameter label for the appropriate parameter (e.g.
\code{"log(sigma)"}). If a function itself is specified (e.g. \code{log} or
\code{function(x) log(x)}) then \code{"t"} is used in the new parameter
label to indicate that the parameter is transformed (e.g.
\code{"t(sigma)"}).}

\item{...}{Currently ignored.}

\item{size, alpha}{Passed to \code{\link[ggplot2]{geom_point}}.}
}
\value{
For \code{mcmc_scatter}, a ggplot object that can be further
  customized using the \pkg{ggplot2} package. For \code{mcmc_pairs}, a
  \code{\link[GGally]{ggpairs}} object.
}
\description{
Scatterplots of MCMC draws. See the \strong{Plot Descriptions} section,
below, for details.
}
\section{Plot Descriptions}{

\describe{
  \item{\code{mcmc_scatter}}{
   Bivariate scatterplot of posterior draws (for two parameters).
  }
  \item{\code{mcmc_pairs}}{
  Scatterplot matrix with histograms along the diagonal.
  }
}
}
\examples{
# some parameter draws to use for demonstration
x <- example_mcmc_draws(params = 6)
dimnames(x)

# scatterplot of alpha vs log(sigma)
(p <- mcmc_scatter(x, pars = c("alpha", "sigma"),
                  trans = list(sigma = "log")))

# add ellipse
p + ggplot2::stat_ellipse(level = 0.9, color = "darkgray", size = 2)

set_color_scheme("purple")
mcmc_scatter(x, pars = c("beta[1]", "beta[4]"))
mcmc_scatter(x, regex = "beta\\\\[[1,4]", alpha = 0.33, size = 3.5)

# pairs plot with histograms along the diagonal
set_color_scheme("mix-green-blue")
mcmc_pairs(x, pars = c("alpha", "sigma", "beta[3]"), alpha = 0.25)

}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other MCMC: \code{\link{MCMC-combos}},
  \code{\link{MCMC-diagnostics}},
  \code{\link{MCMC-distributions}},
  \code{\link{MCMC-intervals}}, \code{\link{MCMC-nuts}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-traces}}
}

