% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-traces.R
\name{MCMC-traces}
\alias{MCMC-traces}
\alias{mcmc_trace}
\alias{mcmc_trace_highlight}
\title{Traceplot (time series plot) of MCMC draws}
\usage{
mcmc_trace(x, pars = character(), regex_pars = character(),
  transformations = list(), n_warmup = 0, inc_warmup = n_warmup > 0,
  window = NULL, size = NULL, facet_args = list(), ...)

mcmc_trace_highlight(x, pars = character(), regex_pars = character(),
  transformations = list(), n_warmup = 0, inc_warmup = n_warmup > 0,
  window = NULL, size = NULL, facet_args = list(), highlight = 1, ...)
}
\arguments{
\item{x}{A 3-D array, matrix, or data frame of posterior draws. If \code{x}
is a 3-D array then the dimensions must be \code{[Iteration, Chain,
Parameter]} in that order. If there is only one Markov chain or if all
chains have already been merged, then \code{x} can be a matrix or
data.frame with a column for each parameter.}

\item{pars}{An optional character vector of parameter names. If neither
\code{pars} nor \code{regex_pars} is specified then the default is to use
\emph{all} parameters.}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}. If neither \code{pars} nor \code{regex_pars} is specified
then the default is to use \emph{all} parameters.}

\item{transformations}{A named list specifying transformations to apply
to parameters. The name of each list element should be a parameter name and
the content of each list element should be a function (or any item to match
as a function via \code{\link{match.fun}}, e.g. a string naming a
function).}

\item{n_warmup}{An integer; the number of warmup iterations included in
\code{x}. The default is \code{n_warmup = 0}, i.e. to assume no warmup
iterations are included.}

\item{inc_warmup}{A logical value indicating if warmup iterations should be
included in the plot if \code{n_warmup > 0}. If \code{inc_warmup} is
\code{TRUE} then the background for warmup iterations is shaded gray.}

\item{window}{An integer vector of length two specifying the limits of a
range of iterations to display.}

\item{size}{An optional value to override the default line size (if calling
\code{mcmc_trace}) or the default point size (if calling
\code{mcmc_trace_highlight}).}

\item{facet_args}{Arguments (other than \code{facets}) passed to
\code{\link[ggplot2]{facet_wrap}} to control faceting.}

\item{...}{Currently ignored.}

\item{highlight}{An integer specifying one of the chains that will be
more visible than the others in the plot.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Traceplot (time series plot) of MCMC draws
}
\section{Plot Descriptions}{

\describe{
  \item{\code{mcmc_trace}}{
   Standard traceplot of MCMC draws.
  }
  \item{\code{mcmc_trace_highlight}}{
   Traces are plotted using points rather than lines and the opacity of all
   chains but one is reduced.
  }
}
}
\seealso{
Other MCMC: \code{\link{MCMC-distributions}},
  \code{\link{MCMC-overview}}, \code{\link{MCMC-pairs}}
}

