% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_dens.R
\name{mcmc_dens}
\alias{mcmc_dens}
\alias{mcmc_dens.array}
\alias{mcmc_dens.data.frame}
\alias{mcmc_dens.matrix}
\title{Kernel density plot of posterior draws}
\usage{
mcmc_dens(x, pars = NULL, regex_pars = NULL, by_chain = FALSE,
  transform = NULL, ...)

\method{mcmc_dens}{array}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, transform = NULL, ...)

\method{mcmc_dens}{matrix}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, transform = NULL, ...)

\method{mcmc_dens}{data.frame}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, transform = NULL, ...)
}
\arguments{
\item{x}{Posterior draws.}

\item{pars}{Parameter names.}

\item{regex_pars}{Regular expression.}

\item{by_chain}{Separate by chain or merge chains?}

\item{...}{Passed to \code{\link[ggplot2]{facet_wrap}} or
\code{\link[ggplot2]{facet_grid}} (e.g. \code{scales}).}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Kernel density plot of posterior draws
}
\seealso{
Other MCMC: \code{\link{mcmc_hist}}
}

