% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-hist.R
\name{mcmc_hist}
\alias{mcmc_hist}
\alias{mcmc_hist.array}
\alias{mcmc_hist.data.frame}
\alias{mcmc_hist.matrix}
\title{Histogram of posterior draws}
\usage{
mcmc_hist(x, pars = NULL, regex_pars = NULL, by_chain = FALSE,
  transform = NULL, binwidth = NULL, facet_args = list(), ...)

\method{mcmc_hist}{array}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, binwidth = NULL, transform = NULL,
  facet_args = list(), ...)

\method{mcmc_hist}{matrix}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, binwidth = NULL, transform = NULL,
  facet_args = list(), ...)

\method{mcmc_hist}{data.frame}(x, pars = NULL, regex_pars = NULL,
  by_chain = FALSE, binwidth = NULL, transform = NULL,
  facet_args = list(), ...)
}
\arguments{
\item{x}{Posterior draws.}

\item{pars}{An optional character vector of parameter names. If not specified
then the first parameter is used.}

\item{by_chain}{Separate by chain or merge chains?}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{facet_args}{Arguments (other than \code{facets}) passed to
\code{\link[ggplot2]{facet_wrap}} (if \code{by_chain} is \code{FALSE}) or
\code{\link[ggplot2]{facet_grid}} (if \code{by_chain} is \code{TRUE}) to
control faceting.}

\item{...}{Currently unused.}

\item{pars}{An optional \link[=grep]{regular expression} to use for parameter
selection. Can be specified instead of \code{pars} or in addition to
\code{pars}.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Histogram of posterior draws
}
\seealso{
Other MCMC: \code{\link{mcmc_dens}},
  \code{\link{mcmc_trace}}
}

