% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-hist.R
\name{mcmc_hist}
\alias{MCMC-hist}
\alias{mcmc_hist}
\title{Histogram of posterior draws}
\usage{
mcmc_hist(x, pars = NULL, regex_pars = NULL, by_chain = FALSE,
  binwidth = NULL, transformations = list(), facet_args = list(), ...)
}
\arguments{
\item{x}{Posterior draws.}

\item{pars}{An optional character vector of parameter names. If not specified
then the first parameter is used.}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}.}

\item{by_chain}{A logical value indicating whether to plot each chain
separately (\code{TRUE}) or merge the chains (\code{FALSE}, the default).}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{transformations}{A named list specifying transformations to apply
to parameters. The name of each list element should be a parameter name and
the content of each list element should be a function (or any item to match
as a function via \code{\link{match.fun}}, e.g. a string naming a
function).}

\item{facet_args}{Arguments (other than \code{facets}) passed to
\code{\link[ggplot2]{facet_wrap}} (if \code{by_chain} is \code{FALSE}) or
\code{\link[ggplot2]{facet_grid}} (if \code{by_chain} is \code{TRUE}) to
control faceting.}

\item{...}{Currently ignored.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Histogram of posterior draws
}
\seealso{
Other MCMC: \code{\link{MCMC-overview}},
  \code{\link{mcmc_dens}}, \code{\link{mcmc_pairs}},
  \code{\link{mcmc_trace}}
}

