% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-trace.R
\name{mcmc_trace}
\alias{MCMC-trace}
\alias{mcmc_trace}
\title{Traceplot (time series plot) of MCMC draws}
\usage{
mcmc_trace(x, pars = NULL, regex_pars = NULL, n_warmup = 0,
  inc_warmup = n_warmup > 0, window = NULL, highlight_chain = NULL,
  style = c("line", "point"), size = NULL, facet_args = list(), ...)
}
\arguments{
\item{x}{Posterior draws.}

\item{pars}{An optional character vector of parameter names. If not specified
then the first parameter is used.}

\item{regex_pars}{An optional \link[=grep]{regular expression} to use for
parameter selection. Can be specified instead of \code{pars} or in addition
to \code{pars}.}

\item{n_warmup}{The number of warmup iterations included in \code{x}.}

\item{inc_warmup}{Include warmup iterations in the plot?}

\item{window}{Integer vector of length two specifying the limits of a range
of iterations to display.}

\item{highlight_chain}{If not \code{NULL}, an integer specifying one of the
chains that will be more visible than the others in the plot.}

\item{style}{Either \code{"points"} or \code{"lines"}.}

\item{size}{An optional value to override \code{\link[ggplot2]{geom_line}}'s
default line size.}

\item{facet_args}{Arguments (other than \code{facets}) passed to
\code{\link[ggplot2]{facet_wrap}} to control faceting.}

\item{...}{Currently ignored.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Traceplot (time series plot) of MCMC draws
}
\seealso{
Other MCMC: \code{\link{MCMC-overview}},
  \code{\link{mcmc_dens}}, \code{\link{mcmc_hist}},
  \code{\link{mcmc_pairs}}
}

