% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc-calibration.R
\name{PPC-calibration}
\alias{PPC-calibration}
\alias{ppc_calibration_overlay}
\alias{ppc_calibration_overlay_grouped}
\alias{ppc_calibration}
\alias{ppc_calibration_grouped}
\alias{ppc_loo_calibration}
\alias{ppc_loo_calibration_grouped}
\alias{ppc_calibration_data}
\alias{ppc_calibration_interval_data}
\alias{ppc_calibration_interval_data_grouped}
\title{PPC calibration}
\usage{
ppc_calibration_overlay(
  y,
  prep,
  ...,
  prob = NULL,
  linewidth = 0.25,
  alpha = 0.2
)

ppc_calibration_overlay_grouped(
  y,
  prep,
  group,
  ...,
  prob = NULL,
  linewidth = 0.25,
  alpha = 0.2
)

ppc_calibration(
  y,
  prep = NULL,
  yrep = NULL,
  prob = 0.95,
  interval = c("confidence", "consistency"),
  help_text = TRUE,
  B = 200,
  show_mean = TRUE,
  show_qdots = TRUE,
  qdots_quantiles = 100,
  ...,
  linewidth = 1,
  alpha = 0.1
)

ppc_calibration_grouped(
  y,
  yrep = NULL,
  prep = NULL,
  group,
  prob = 0.95,
  interval = c("confidence", "consistency"),
  help_text = TRUE,
  B = 200,
  show_mean = TRUE,
  show_qdots = TRUE,
  qdots_quantiles = 100,
  ...,
  linewidth = 1,
  alpha = 0.1
)

ppc_loo_calibration(
  y,
  yrep,
  lw = NULL,
  psis_object = NULL,
  prob = 0.95,
  interval = c("confidence", "consistency"),
  help_text = TRUE,
  B = 200,
  show_mean = TRUE,
  show_qdots = TRUE,
  qdots_quantiles = 100,
  ...,
  linewidth = 1,
  alpha = 0.1
)

ppc_loo_calibration_grouped(
  y,
  yrep,
  lw = NULL,
  psis_object = NULL,
  group,
  prob = 0.95,
  interval = c("confidence", "consistency"),
  help_text = TRUE,
  B = 200,
  show_mean = TRUE,
  show_qdots = TRUE,
  qdots_quantiles = 100,
  ...,
  linewidth = 1,
  alpha = 0.1
)

ppc_calibration_data(y, prep, group = NULL)

ppc_calibration_interval_data(
  y,
  prep = NULL,
  yrep = NULL,
  prob = 0.95,
  interval = c("confidence", "consistency"),
  B = 200
)

ppc_calibration_interval_data_grouped(
  y,
  prep = NULL,
  yrep = NULL,
  group,
  prob = 0.95,
  interval = c("confidence", "consistency"),
  B = 200
)
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{prep}{For \code{ppc_calibration()}, \code{ppc_calibration_grouped()},
\code{ppc_calibration_overlay()}, and \code{ppc_calibration_overlay_grouped()},
an \code{S} by \code{N} matrix of predicted probabilities in \verb{[0, 1]}, where \code{S} is
the number of draws and \code{N = length(y)}.}

\item{...}{Currently unused.}

\item{prob}{Probability mass to include in the plotted interval.}

\item{linewidth, alpha}{Arguments passed to geoms controlling line width and
opacity.}

\item{group}{A grouping variable of the same length as \code{y}.
Will be coerced to \link[base:factor]{factor} if not already a factor.
Each value in \code{group} is interpreted as the group level pertaining
to the corresponding observation.}

\item{yrep}{An \code{S} by \code{N} matrix of draws from the posterior (or prior)
predictive distribution, or a \code{\link[posterior:draws]{posterior::draws}} object. The number of
rows, \code{S}, is the size of the posterior (or prior) sample used to generate
\code{yrep}. The number of columns, \code{N} is the number of predicted observations
(\code{length(y)}). The columns of \code{yrep} should be in the same order as the
data points in \code{y} for the plots to make sense. See the \strong{Details} and
\strong{Plot Descriptions} sections for additional advice specific to particular
plots.}

\item{interval}{For \code{ppc_calibration()}, \code{ppc_calibration_grouped()},
\code{ppc_loo_calibration()}, and \code{ppc_loo_calibration_grouped()}, choose
\code{"confidence"} to use uncertainty in the estimated calibration curve, or
\code{"consistency"} to compare to replicated outcomes.}

\item{help_text}{For \code{ppc_calibration()} and \code{ppc_calibration_grouped()},
if \code{TRUE} (default) display a math-style label in the plot indicating the
interval as \code{CI} with the selected \code{prob} and \code{interval} values.}

\item{B}{For calibration plots that use \code{yrep} with \code{interval = "confidence"},
the number of bootstrap samples.}

\item{show_mean}{For calibration interval plots, if \code{TRUE} (default), draw
the estimated calibration curve.}

\item{show_qdots}{For \code{ppc_calibration()}, \code{ppc_calibration_grouped()},
\code{ppc_loo_calibration()}, and \code{ppc_loo_calibration_grouped()}, if \code{TRUE}
add a quantile dot plot at the bottom of the panel to show the marginal
distribution of predicted probabilities.}

\item{qdots_quantiles}{For \code{ppc_calibration()}, \code{ppc_calibration_grouped()},
\code{ppc_loo_calibration()}, and \code{ppc_loo_calibration_grouped()}, positive
integer passed to
the \code{quantiles} argument of \code{ggdist::stat_dots()} when
\code{show_qdots = TRUE}.}

\item{lw}{For \code{ppc_loo_calibration()} and \code{ppc_loo_calibration_grouped()},
a matrix of log weights with the same dimensions as \code{yrep}. Either
\code{psis_object} or \code{lw} has to be specified.}

\item{psis_object}{For \code{ppc_loo_calibration()} and \code{ppc_loo_calibration_grouped()},
an object of class \code{"psis"} that is created when the \code{loo()} function calls
\code{psis()} internally to do the PSIS procedure. Either \code{psis_object} or \code{lw}
has to be specified.}
}
\value{
The plotting functions return a ggplot object that can be further
customized using the \strong{ggplot2} package. The functions with suffix
\verb{_data()} return the data that would have been drawn by the plotting
function.
}
\description{
Assess the calibration of the predictions, or predictive probabilites in relation to
binary observations.
See the \strong{Plot Descriptions} section, below, for details.
}
\section{Plot Descriptions}{

\describe{
\item{\code{ppc_calibration()},\code{ppc_calibration_grouped()}}{
PAV-adjusted calibration plots showing the relationship between the
predicted event probabilities and the conditional event probabilities.
The \code{interval} parameter controls whether confidence intervals, or
consistency intervals are computed.},
\item{\code{ppc_calibration_overlay()},\code{ppc_calibration_overlay_grouped()}}{
Overlay plots showing posterior samples of PAV-adjusted calibration
curves.
},
\item{\code{ppc_loo_calibration()},\code{ppc_loo_calibration_grouped()}}{
PAV-adjusted calibration plots to assess the calibration of the
leave-one-out (LOO) predictive probabilities, computed by resampling each
observation's posterior predictive draws using the corresponding LOO
importance weights.
}
}
}

\examples{
color_scheme_set("brightblue")

# Make an example dataset of binary observations
ymin <- range(example_y_data(), example_yrep_draws())[1]
ymax <- range(example_y_data(), example_yrep_draws())[2]
y <- rbinom(length(example_y_data()), 1, (example_y_data() - ymin) / (ymax - ymin))
prep <- (example_yrep_draws() - ymin) / (ymax - ymin)

ppc_calibration_overlay(y, prep[1:50, ])

# Compare confidence vs consistency intervals
ppc_calibration(y, prep, interval = "confidence")
ppc_calibration(y, prep, interval = "consistency")
}
\seealso{
Other PPCs: 
\code{\link{PPC-censoring}},
\code{\link{PPC-discrete}},
\code{\link{PPC-distributions}},
\code{\link{PPC-errors}},
\code{\link{PPC-intervals}},
\code{\link{PPC-loo}},
\code{\link{PPC-overview}},
\code{\link{PPC-scatterplots}},
\code{\link{PPC-test-statistics}}
}
\concept{PPCs}
