% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc-calibration.R
\name{PPC-calibration}
\alias{PPC-calibration}
\alias{ppc_calibration_overlay}
\alias{ppc_calibration_overlay_grouped}
\alias{ppc_calibration}
\alias{ppc_calibration_grouped}
\alias{ppc_loo_calibration}
\alias{ppc_loo_calibration_grouped}
\title{Assess the calibration of the predictive distributions \code{yrep} in relation to
the data `y'.
See the \strong{Plot Descriptions} section, below, for details.}
\usage{
ppc_calibration_overlay(y, prep, ..., linewidth = 0.25, alpha = 0.5)

ppc_calibration_overlay_grouped(
  y,
  prep,
  group,
  ...,
  linewidth = 0.25,
  alpha = 0.7
)

ppc_calibration(y, prep, prob = 0.95, ..., linewidth = 0.5, alpha = 0.7)

ppc_calibration_grouped(
  y,
  prep,
  group,
  prob = 0.95,
  ...,
  linewidth = 0.5,
  alpha = 0.7
)

ppc_loo_calibration(y, yrep, lw, ..., linewidth = 0.25, alpha = 0.7)

ppc_loo_calibration_grouped(
  y,
  yrep,
  group,
  lw,
  ...,
  linewidth = 0.25,
  alpha = 0.7
)
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{group}{A grouping variable of the same length as \code{y}.
Will be coerced to \link[base:factor]{factor} if not already a factor.
Each value in \code{group} is interpreted as the group level pertaining
to the corresponding observation.}

\item{yrep}{An \code{S} by \code{N} matrix of draws from the posterior (or prior)
predictive distribution. The number of rows, \code{S}, is the size of the
posterior (or prior) sample used to generate \code{yrep}. The number of columns,
\code{N} is the number of predicted observations (\code{length(y)}). The columns of
\code{yrep} should be in the same order as the data points in \code{y} for the plots
to make sense. See the \strong{Details} and \strong{Plot Descriptions} sections for
additional advice specific to particular plots.}
}
\value{
The plotting functions return a ggplot object that can be further
customized using the \strong{ggplot2} package. The functions with suffix
\verb{_data()} return the data that would have been drawn by the plotting
function.
}
\description{
Assess the calibration of the predictive distributions \code{yrep} in relation to
the data `y'.
See the \strong{Plot Descriptions} section, below, for details.
}
\section{Plot Descriptions}{

\describe{
\item{\code{ppc_calibration()},\code{ppc_calibration_grouped()}}{

},
\item{\code{ppc_calibration_overlay()},\code{ppc_calibration_overlay_grouped()}}{

},
\item{\code{ppc_loo_calibration()},\code{ppc_loo_calibration_grouped()}}{

}
}
}

\examples{
color_scheme_set("brightblue")

# Make an example dataset of binary observations
ymin <- range(example_y_data(), example_yrep_draws())[1]
ymax <- range(example_y_data(), example_yrep_draws())[2]
y <- rbinom(length(example_y_data()), 1, (example_y_data() - ymin) / (ymax - ymin))
prep <- (example_yrep_draws() - ymin) / (ymax - ymin)

ppc_calibration_overlay(y, prep[1:50, ])
}
\seealso{
Other PPCs: 
\code{\link{PPC-censoring}},
\code{\link{PPC-discrete}},
\code{\link{PPC-distributions}},
\code{\link{PPC-errors}},
\code{\link{PPC-intervals}},
\code{\link{PPC-loo}},
\code{\link{PPC-overview}},
\code{\link{PPC-scatterplots}},
\code{\link{PPC-test-statistics}}
}
\concept{PPCs}
