% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc-distributions.R
\name{PPC-distributions}
\alias{PPC-distributions}
\alias{ppc_dens}
\alias{ppc_dens_overlay}
\alias{ppc_ecdf_overlay}
\alias{ppc_hist}
\alias{ppc_violin_grouped}
\title{PPC distributions}
\usage{
ppc_hist(y, yrep, ..., binwidth = NULL)

ppc_dens(y, yrep, ..., trim = FALSE)

ppc_dens_overlay(y, yrep, ..., size = 0.25, alpha = 0.7, trim = FALSE)

ppc_ecdf_overlay(y, yrep, ..., size = 0.25, alpha = 0.7, pad = TRUE)

ppc_violin_grouped(y, yrep, group, ..., probs = c(0.1, 0.5, 0.9), alpha = 1,
  size = 1)
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See \strong{Details} for additional
instructions.}

\item{...}{Currently unused.}

\item{binwidth}{An optional value used as the \code{binwidth} argument to
\code{\link[ggplot2]{geom_histogram}} to override the default binwidth.}

\item{trim}{A logical scalar passed to \code{\link[ggplot2]{geom_density}}.}

\item{size, alpha}{Passed to the appropriate geom to control the appearance of
the \code{yrep} distributions.}

\item{pad}{For \code{ppc_ecdf_overlay}, a logical scalar passed to
\code{\link[ggplot2]{stat_ecdf}}.}

\item{group}{A grouping variable (a vector or factor) the same length as
\code{y}. Each value in \code{group} is interpreted as the group level
pertaining to the corresponding value of \code{y}.}

\item{probs}{A numeric vector passed to \code{\link[ggplot2]{geom_violin}}'s
\code{draw_quantiles} argument to specify at which quantiles to draw
horizontal lines. Set to \code{NULL} to remove the lines.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Compare the empirical distribution of the data \code{y} to the distributions
of simulated/replicated data \code{yrep} from the posterior predictive
distribution. See the \strong{Plot Descriptions} section, below,
for details.
}
\details{
For Binomial data, the plots will typically be most useful if
  \code{y} and \code{yrep} contain the "success" proportions (not discrete
  "success" or "failure" counts).
}
\section{Plot Descriptions}{

\describe{
  \item{\code{ppc_dens_overlay}}{
   Kernel density estimates of each dataset (row) in \code{yrep} are
   overlaid, with the distribution of \code{y} itself on top (and in a darker
   shade).
  }
  \item{\code{ppc_ecdf_overlay}}{
   Empirical CDF estimates of each dataset (row) in \code{yrep} are overlaid,
   with the distribution of \code{y} itself on top (and in a darker shade).
  }
  \item{\code{ppc_hist}}{
   A separate histogram is plotted for \code{y} and each dataset (row) in
   \code{yrep}. For this plot \code{yrep} should therefore contain only a
   small number of rows. See the \strong{Examples} section.
  }
  \item{\code{ppc_dens}}{
   The same as \code{ppc_hist} but kernel density estimates are plotted
   instead of histograms. These plots often look like wizard hats.
  }
  \item{\code{ppc_violin_grouped}}{
   The density estimate of \code{yrep} within each level of a grouping
   variable is plotted as a violin with horizontal lines at
   notable quantiles. The points in \code{y} corresponding to
   each grouping level are then overlaid on top of the violins.
  }
}
}
\examples{
color_scheme_set("brightblue")
y <- example_y_data()
yrep <- example_yrep_draws()
dim(yrep)
ppc_dens_overlay(y, yrep[1:50, ])
ppc_ecdf_overlay(y, yrep[sample(nrow(yrep), 25), ])

# for ppc_hist, definitely use a subset yrep rows so only
# a few (instead of nrow(yrep)) histograms are plotted
ppc_hist(y, yrep[1:8, ])

# wizard hat plot
color_scheme_set("blue")
ppc_dens(y, yrep[200:202, ])

color_scheme_set("gray")
group <- example_group_data()
ppc_violin_grouped(y, yrep, group, size = 1.5)
ppc_violin_grouped(y, yrep, group, alpha = 0)

}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
Other PPCs: \code{\link{PPC-overview}},
  \code{\link{PPC-residuals}},
  \code{\link{PPC-scatterplots}},
  \code{\link{PPC-test-statistics}},
  \code{\link{PPC-time-series}}, \code{\link{PPC-vs-x}}
}

