% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc-distributions.R
\name{PPC-distributions}
\alias{PPC-distributions}
\alias{ppc_dens}
\alias{ppc_dens_overlay}
\alias{ppc_hist}
\alias{ppc_violin_grouped}
\title{PPC distributions}
\usage{
ppc_hist(y, yrep, ..., binwidth = NULL)

ppc_dens(y, yrep, ...)

ppc_dens_overlay(y, yrep, ...)

ppc_violin_grouped(y, yrep, group, ..., probs = c(0.1, 0.5, 0.9))
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See \strong{Details} for additional
instructions.}

\item{...}{Currently unused.}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{group}{A grouping variable (a vector or factor) the same length as
\code{y}. Each value in \code{group} is interpreted as the group level
pertaining to the corresponding value of \code{y}.}

\item{probs}{A numeric vector passed to \code{\link[ggplot2]{geom_violin}}'s
\code{draw_quantiles} argument to specify at which quantiles to draw
horizontal lines. Set to \code{NULL} to remove the lines.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Compare the empirical distribution of the data \code{y} to the distributions
of simulated/replicated data \code{yrep} from the posterior predictive
distribution.
}
\details{
For Binomial data, the plots will typically be most useful if
  \code{y} and \code{yrep} contain the "success" proportions (not discrete
  "success" or "failure" counts).
}
\section{Plot Descriptions}{

\describe{
  \item{\code{ppc_dens_overlay}}{
   The kernel density estimate for \code{y} is plotted with the area under
   the curve shaded. The density estimates of each dataset (row) in
   \code{yrep} are then overlaid but unshaded.
  }
  \item{\code{ppc_hist}}{
   A separate histogram is plotted for \code{y} and each dataset (row) in
   \code{yrep}. For this plot \code{yrep} should therefore contain only a
   small number of rows.
  }
  \item{\code{ppc_dens}}{
   The same as \code{ppc_hist} but kernel density estimates are plotted
   instead of histograms.
  }
  \item{\code{ppc_violin_grouped}}{
   The density estimate of \code{yrep} within each level of a grouping
   variable is plotted as a violin with horizontal lines at
   notable quantiles. The points in \code{y} corresponding to
   each grouping level are then overlaid on top of the violins.
  }
}
}
\examples{
y <- rnorm(100)
yrep <- matrix(rnorm(2500), ncol = 100)
ppc_dens_overlay(y, yrep)
ppc_hist(y, yrep[1:8, ])

group <- gl(4, 25, labels = LETTERS[1:4])
ppc_violin_grouped(y, yrep, group)

}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other PPCs: \code{\link{PPC-overview}},
  \code{\link{PPC-residuals}},
  \code{\link{PPC-scatterplots}},
  \code{\link{PPC-test-statistics}},
  \code{\link{PPC-time-series}}
}

