% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc-residuals.R
\name{PPC-residuals}
\alias{PPC-residuals}
\alias{ppc_resid_binned}
\alias{ppc_resid_hist}
\alias{ppc_resid_scatter}
\title{PPC residuals}
\usage{
ppc_resid_hist(y, yrep, ..., binwidth = NULL)

ppc_resid_scatter(y, yrep, ..., size = 2.5, alpha = 0.8)

ppc_resid_binned(y, Ey, ...)
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See \strong{Details} for additional
instructions.}

\item{...}{Currently unused.}

\item{binwidth}{An optional value used as the \code{binwidth} argument to
\code{\link[ggplot2]{geom_histogram}} to override the default binwidth.}

\item{size, alpha}{Arguments passed to \code{\link[ggplot2]{geom_point}} to
control the appearance of scatterplot points.}

\item{Ey}{A matrix of posterior draws of the linear predictor transformed by
the inverse-link function.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
\code{ppc_resid_hist} plots histograms of residuals computed from \code{y}
and simulated datasets \code{yrep}. \code{ppc_resid_scatter} plots
scatterplots of residuals vs. \code{y}. For binomial data,
\code{ppc_resid_binned} generates binned residual plots (similar to
\code{\link[arm]{binnedplot}}) from \code{y} and the posterior draws of the
linear predictor transformed by the inverse-link function. See the
\strong{Details} and \strong{Plot Descriptions} sections, below.
}
\details{
These plots compute and plot residuals for each row of the matrix \code{yrep}
(or for \code{ppc_resid_binned} the matrix \code{Ey}), so it is usually a
good idea for \code{yrep} and \code{Ey} to contain only a small number of
draws (rows). See \strong{Examples}, below.

For binomial and Bernoulli data the \code{ppc_resid_binned} function
can be used to generate binned residual plots. Bernoulli data can be input
as a vector of 0s and 1s, whereas for binomial data \code{y} should be a
vector of "success" proportions (not a matrix of "success" and "failure"
counts).
}
\section{Plot descriptions}{

\describe{
  \item{\code{ppc_resid_hist}}{
   A separate histogram is plotted for the residuals computed from \code{y}
   and each dataset (row) in \code{yrep}. For this plot
   \code{yrep} should have only a small number of rows.
  }
  \item{\code{ppc_resid_scatter}}{
   A separate scatterplot is displayed for \code{y} vs. the residuals
   computed from \code{y} and each dataset (row) in \code{yrep}. For this
   plot \code{yrep} should have only a small number of rows.
  }
  \item{\code{ppc_resid_binned}}{
   Intended for use with binomial data. A separate binned residual plot
   (similar to \code{\link[arm]{binnedplot}}) is generated for each dataset
   (row) in \code{Ey}, the posterior draws of the linear predictor
   transformed by the inverse-link function. For this plot \code{Ey} should
   have only a small number of rows.
  }
}
}
\examples{
y <- example_y_data()
yrep <- example_yrep_draws()
ppc_resid_hist(y, yrep[1:3, ])
ppc_resid_scatter(y, yrep[10:14, ])

}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other PPCs: \code{\link{PPC-distributions}},
  \code{\link{PPC-overview}},
  \code{\link{PPC-scatterplots}},
  \code{\link{PPC-test-statistics}},
  \code{\link{PPC-time-series}}, \code{\link{PPC-vs-x}}
}

