% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc-rootograms.R
\name{PPC-rootograms}
\alias{PPC-rootograms}
\alias{ppc_rootogram}
\title{PPC rootograms}
\usage{
ppc_rootogram(y, yrep, style = c("standing", "hanging", "suspended"), ...,
  prob = 0.9, size = 1)
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See \strong{Details} for additional
instructions.}

\item{style}{The rootogram style. The options are \code{"standing"},
\code{"hanging"}, and \code{"suspended"}. See the \strong{Plot
Descriptions} section, below, for details on the different styles.}

\item{...}{Currently unused.}

\item{prob}{The probability mass to include in the uncertainty interval
around the square roots of the expected counts. Defaults to \code{0.9}. Set
\code{prob=0} to remove the uncertainty interval.}

\item{size}{Passed to \code{\link[ggplot2]{geom_line}}.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Rootograms allow for diagnosing problems in count data models such as
overdispersion or excess zeros. They consist of a histogram of \code{y} with
the expected counts based on \code{yrep} overlaid as a line along with
uncertainty intervals. The histogram style may be adjusted to focus on
different aspects of the data. The y-axis represents the square roots of the
counts to approximately adjust for scale differences and thus ease comparison
between observed and expected counts. For a detailed introduction to
rootograms see Kleiber and Zeileis (2016).
}
\details{
For rootograms, the observations \code{y} and predictons \code{yrep}
  must be counts. \pkg{bayesplot} will validate that both \code{y} and
  \code{yrep} contain only non-negative integer values, although they need
  not be integers in the sense of \R's \code{\link{integer}} type.
}
\section{Plot Descriptions}{

\describe{
\item{\code{ppc_rootogram}}{}
 \itemize{
  \item \emph{Standing}: basic histogram of observed counts with curve
  showing expected counts.
  \item \emph{Hanging}: observed counts counts hanging from the curve
  representing expected counts.
  \item \emph{Suspended}: histogram of the differences between expected and
  observed counts.
 }
\strong{All of these are plotted on the square root scale}. See Kleiber and
Zeileis (2016) for advice on interpreting rootograms and selecting among the
different styles.
}
}

\examples{
y <- rpois(100, 20)
yrep <- matrix(rpois(10000, 20), ncol = 100)

ppc_rootogram(y, yrep)
ppc_rootogram(y, yrep, style = "hanging", prob = 0.8)
ppc_rootogram(y, yrep, style = "suspended")

}
\references{
Kleiber, C. and Zeileis, A. (2016). Visualizing count data regressions using
rootograms. \emph{The American Statistician}. 70(3): 296--303.
\url{https://arxiv.org/abs/1605.01311}.
}
\seealso{
Other PPCs: \code{\link{PPC-distributions}},
  \code{\link{PPC-errors}}, \code{\link{PPC-intervals}},
  \code{\link{PPC-overview}},
  \code{\link{PPC-scatterplots}},
  \code{\link{PPC-test-statistics}}
}
