% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc-rootograms.R
\name{PPC-rootograms}
\alias{PPC-rootograms}
\alias{ppc_rootogram}
\title{PPC rootograms}
\usage{
ppc_rootogram(y, yrep, style = c("standing", "suspended"), ..., prob = 0.9,
  size = 1)
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See \strong{Details} for additional
instructions.}

\item{style}{Rootogram style. Either \code{"standing"} or \code{"suspended"}.}

\item{...}{Currently unused.}

\item{prob}{The probability mass to include in the uncertainty interval
around the expected (root) counts. Defaults to \code{0.9}.}

\item{size}{Passed to \code{\link[ggplot2]{geom_line}}.}
}
\description{
FIXME: add description
}
\seealso{
Other PPCs: \code{\link{PPC-distributions}},
  \code{\link{PPC-errors}}, \code{\link{PPC-intervals}},
  \code{\link{PPC-overview}},
  \code{\link{PPC-scatterplots}},
  \code{\link{PPC-test-statistics}}
}

