% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-statistics.R
\name{PPC-test-statistics}
\alias{PPC-test-statistics}
\alias{ppc_stat}
\alias{ppc_stat_2d}
\alias{ppc_stat_grouped}
\title{PPC test statistics}
\usage{
ppc_stat(y, yrep, stat = "mean", ..., binwidth = NULL)

ppc_stat_grouped(y, yrep, group, stat = "mean", ..., binwidth = NULL)

ppc_stat_2d(y, yrep, stat = c("mean", "sd"), ...)
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See \strong{Details} for additional
instructions.}

\item{stat}{A character vector of function names of length 1 (for
\code{ppc_stat} and \code{ppc_stat_grouped}) or length 2 (for
\code{ppc_stat_2d}). The function(s) should take a vector input and return
a scalar test statistic.}

\item{...}{Currently unused.}

\item{binwidth}{An optional value to override
\code{\link[ggplot2]{geom_histogram}}'s default binwidth.}

\item{group}{A grouping variable (a vector or factor) the same length as
\code{y}. Each value in \code{group} is interpreted as the group level
pertaining to the corresponding value of \code{y}.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
The distribution of a test statistic \code{T(yrep)}, or a
pair of test statistics, over the simulated datasets in \code{yrep},
compared to the observed value \code{T(y)}.
}
\details{
For Binomial data, the plots will typically be most useful if
  \code{y} and \code{yrep} contain the "success" proportions (not discrete
  "success" or "failure" counts).
}
\section{Plot Descriptions}{

\describe{
  \item{\code{ppc_stat}}{
   A histogram of the distribution of a test statistic computed by applying
   \code{stat} to each dataset (row) in \code{yrep}. The value of the
   statistic in the observed data, \code{stat(y)}, is overlaid as a vertical
   line.
  }
  \item{\code{ppc_stat_grouped}}{
   The same as \code{ppc_stat}, but a separate plot is generated for
   each level of a grouping variable.
  }
  \item{\code{ppc_stat_2d}}{
   A scatterplot showing the joint distribution of two test statistics
   computed over the datasets (rows) in \code{yrep}. The value of the
   statistics in the observed data is overlaid as large point.
  }
}
}
\examples{
y <- rnorm(30)
yrep <- matrix(rnorm(3000), ncol = 30)
ppc_stat(y, yrep)
ppc_stat(y, yrep, stat = "var", binwidth = 0.1)
ppc_stat_2d(y, yrep)
ppc_stat_2d(y, yrep, stat = c("median", "mean"))

group <- gl(3, 10, labels = LETTERS[1:3])
ppc_stat_grouped(y, yrep, group)

# use your own function to compute test statistics
q25 <- function(y) quantile(y, 0.25)
ppc_stat(y, yrep, stat = "q25")

}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other PPCs: \code{\link{PPC-distributions}},
  \code{\link{PPC-overview}}, \code{\link{PPC-residuals}},
  \code{\link{PPC-scatterplots}},
  \code{\link{PPC-time-series}}
}

