% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_bars.R
\name{ppc_bars}
\alias{ppc_bars}
\alias{ppc_bars_grouped}
\title{Bar plots for discrete outcomes (ordinal, multinomial, count)}
\usage{
ppc_bars(y, yrep, ..., prob = 0.9, width = 0.9, size = 1, fatten = 3)

ppc_bars_grouped(y, yrep, group, facet_args = list(), ..., prob = 0.9,
  width = 0.9, size = 1, fatten = 3)
}
\arguments{
\item{y}{A vector of observations. See \strong{Details}.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}). The columns
of \code{yrep} should be in the same order as the data points in \code{y}
for the plots to make sense. See \strong{Details} for additional
instructions.}

\item{prob}{A value between 0 and 1 indicating the desired probability mass
to include in the \code{yrep} intervals. The default is 0.9.}

\item{width}{Passed to \code{\link[ggplot2]{geom_bar}} to control the bar
width.}

\item{size, fatten}{Passed to \code{\link[ggplot2]{geom_pointrange}} to
control the appearance of the \code{yrep} points and intervals.}

\item{group}{A grouping variable (a vector or factor) the same length as
\code{y}. Each value in \code{group} is interpreted as the group level
pertaining to the corresponding value of \code{y}.}

\item{facet_args}{An optional list of  arguments (other than \code{facets})
passed to \code{\link[ggplot2]{facet_wrap}} to control faceting.}
}
\description{
Bar plots for discrete outcomes (ordinal, multinomial, count). See the
\strong{Plot Descriptions} section below.
}
\details{
For \code{ppc_bars}, the observations \code{y} and predictons
  \code{yrep} must be non-negative integers. \pkg{bayesplot} will validate
  that both \code{y} and \code{yrep} contain only non-negative integer
  values, although they need not be integers in the strict sense of \R's
  \code{\link{integer}} type.
}
\section{Plot Descriptions}{

\describe{
\item{\code{ppc_bars}}{
  Bar plot of a discrete variable \code{y} with \code{yrep} medians and
  uncertainty intervals superimposed on the bars.
}
}
}

