% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppd-distributions.R
\name{PPD-distributions}
\alias{PPD-distributions}
\alias{ppd_data}
\alias{ppd_dens}
\alias{ppd_hist}
\alias{ppd_dens_overlay}
\alias{ppd_boxplot}
\title{PPD distributions}
\usage{
ppd_data(ypred, group = NULL)

ppd_dens(ypred, ..., trim = FALSE, size = 0.5, alpha = 1)

ppd_hist(ypred, ..., binwidth = NULL, breaks = NULL, freq = TRUE)

ppd_dens_overlay(ypred, ..., size = 0.25, alpha = 0.7, trim = FALSE,
  bw = "nrd0", adjust = 1, kernel = "gaussian", n_dens = 1024)

ppd_boxplot(ypred, ..., notch = TRUE, size = 0.5, alpha = 1)
}
\arguments{
\item{ypred}{An \eqn{S} by \eqn{N} matrix of draws from the posterior (or
prior) predictive distribution, where \eqn{S} is the size of the posterior
(or prior) sample (or subset of the posterior sample used to generate
\code{ypred}) and \eqn{N} is the number of predicted observations.}

\item{group}{A grouping variable (a vector or factor) the same length as
\code{y}. Each value in \code{group} is interpreted as the group level
pertaining to the corresponding value of \code{y}.}

\item{...}{Currently unused.}

\item{trim}{A logical scalar passed to \code{\link[ggplot2:geom_density]{ggplot2::geom_density()}}.}

\item{size}{Passed to the appropriate geom to control the appearance of
the \code{yrep} distributions.}

\item{alpha}{Passed to the appropriate geom to control the appearance of
the \code{yrep} distributions.}

\item{binwidth}{Passed to \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}} to override
the default binwidth.}

\item{breaks}{Passed to \code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}} as an
alternative to \code{binwidth}.}

\item{freq}{For histograms, \code{freq=TRUE} (the default) puts count on the
y-axis. Setting \code{freq=FALSE} puts density on the y-axis. (For many
plots the y-axis text is off by default. To view the count or density
labels on the y-axis see the \code{\link[=yaxis_text]{yaxis_text()}} convenience
function.)}

\item{bw}{Optional arguments passed to
\code{\link[stats:density]{stats::density()}} to override default kernel density estimation
parameters. \code{n_dens} defaults to \code{1024}.}

\item{adjust}{Optional arguments passed to
\code{\link[stats:density]{stats::density()}} to override default kernel density estimation
parameters. \code{n_dens} defaults to \code{1024}.}

\item{kernel}{Optional arguments passed to
\code{\link[stats:density]{stats::density()}} to override default kernel density estimation
parameters. \code{n_dens} defaults to \code{1024}.}

\item{n_dens}{Optional arguments passed to
\code{\link[stats:density]{stats::density()}} to override default kernel density estimation
parameters. \code{n_dens} defaults to \code{1024}.}

\item{notch}{A logical scalar passed to \code{\link[ggplot2:geom_boxplot]{ggplot2::geom_boxplot()}}.
Unlike for \code{geom_boxplot()}, the default is \code{notch=TRUE}.}
}
\description{
Plot posterior or prior predictive distributions. Each of these functions
makes the same plot as its corresponding \code{ppc_} function but without plotting
observed data \code{y}. The \strong{Plot Descriptions} section at \link{PPC-distributions}
has details on the individual plots.
}
\examples{
# difference between ppd_dens_overlay and ppc_dens_overlay
preds <- example_yrep_draws()
ppd_dens_overlay(ypred = preds[1:50, ])
ppc_dens_overlay(y = example_y_data(), yrep = preds[1:50, ])

}
\concept{PPDs}
