% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-scheme.R
\name{set_color_scheme}
\alias{get_color_scheme}
\alias{set_color_scheme}
\title{Set or get the color scheme}
\usage{
set_color_scheme(scheme = c("reds", "blues", "greens", "greys", "purples"))

get_color_scheme()
}
\arguments{
\item{scheme}{A string naming the color scheme to use.}
}
\value{
\code{set_color_scheme} has the side effect of setting the color
  scheme used for plotting. It also returns,
  \code{\link[=invisible]{invisibly}}, a list of the hexidecimal color values
  used in \code{scheme}.

\code{get_color_scheme} returns a \code{list} of the hexadecimal
  color values used by the current scheme.
}
\description{
Set or get the color scheme used for plotting.
}
\examples{
get_color_scheme()
y <- rnorm(100)
yrep <- matrix(rnorm(2500), ncol = 100)
ppc_stat_2d(y, yrep)

set_color_scheme("greens")
ppc_stat_2d(y, yrep)

set_color_scheme("purples")
ppc_hist(y, yrep[1:8,])

}

