% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-scheme.R
\name{set_color_scheme}
\alias{get_color_scheme}
\alias{set_color_scheme}
\title{Set or get the color scheme}
\usage{
set_color_scheme(scheme = c("red", "blue", "gray", "green", "pink", "teal"))

get_color_scheme(scheme)
}
\arguments{
\item{scheme}{A string naming a color scheme.}
}
\value{
\code{set_color_scheme} has the side effect of setting the color
  scheme used for plotting. It also returns
  (\code{\link[=invisible]{invisibly}}) a list of the hexidecimal color
  values used in \code{scheme}.

\code{get_color_scheme} returns a \code{list} of the hexadecimal
  color values (without changing the current scheme). If the \code{scheme}
  argument is not specified the returned values correspond to the current
  color scheme.
}
\description{
Set or get the color scheme used for plotting.
}
\examples{
get_color_scheme()

x <- fake_draws()
mcmc_intervals(x)

set_color_scheme("teal")
mcmc_intervals(x)

set_color_scheme("blue")
mcmc_areas(x)

set_color_scheme("pink")
y <- rnorm(100)
yrep <- matrix(rnorm(5e4), ncol = 100)
ppc_stat(y, yrep, stat = "sd") + no_legend()

# lightest color in the 'gray' color scheme
get_color_scheme("gray")$light

}

