% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-params.R
\name{tidy-params}
\alias{tidy-params}
\title{Tidy parameter selection}
\description{
As of version \code{1.7.0}, \strong{bayesplot} allows the \code{pars} argument
for \link[bayesplot:MCMC-overview]{MCMC plots} to be used for so-called 'tidy'
variable selection (in the style of the \strong{dplyr} package). The
\code{\link[dplyr:vars]{vars()}} function is re-exported from \strong{dplyr} for this
purpose.

When using \code{pars} for tidy parameter selection, the \code{regex_pars} argument
is ignored because \strong{bayesplot} supports using \link[tidyselect:select_helpers]{tidyselect helper functions} (\code{starts_with()}, \code{contains()}, etc.)
for the same purpose.
}
\examples{

x <- example_mcmc_draws(params = 6)
dimnames(x)
mcmc_hex(x, pars = vars(alpha, `beta[2]`))
mcmc_dens(x, pars = vars(sigma, contains("beta")))
mcmc_hist(x, pars = vars(-contains("beta")))

\donttest{
if (requireNamespace("rstanarm", quietly = TRUE)) {
  fit <- example("example_model", package = "rstanarm", local=TRUE)$value
  posterior <- as.matrix(fit)
  print(colnames(posterior))

  color_scheme_set("brightblue")

  mcmc_hist(posterior, pars = vars(size, period3))
  mcmc_hist(posterior, pars = vars(size, contains("period")))
  mcmc_hist(posterior, pars = vars(size, contains("period"), -period3),
            facet_args = list(ncol = 2))

  mcmc_intervals(posterior, pars = vars(contains("herd")))
  mcmc_intervals(posterior, pars = vars(contains("herd"), -contains("Sigma")))

  bayesplot_theme_set(ggplot2::theme_dark())
  color_scheme_set("viridisB")
  mcmc_areas_ridges(posterior, pars = vars(starts_with("b[")))

  color_scheme_set("purple")
  mcmc_areas_ridges(posterior, pars = vars(starts_with("b["), -matches("[7-9]")))
}
}


}
