% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-series.R
\name{time-series}
\alias{ppc_ts}
\alias{ppc_ts_grouped}
\alias{time-series}
\title{PPCs for time series}
\usage{
ppc_ts(y, yrep, ..., prob = 0.8)

ppc_ts_grouped(y, yrep, group, ..., prob = 0.8)
}
\arguments{
\item{y}{A vector of observed data. See Details.}

\item{yrep}{An \eqn{S} by \eqn{N} matrix of draws from the posterior
predictive distribution, where \eqn{S} is the size of the posterior sample
(or subset of the posterior sample used to generate \code{yrep}) and
\eqn{N} is the number of observations (the length of \code{y}).
See Details for additional instructions.}

\item{...}{Currently unused.}

\item{prob}{A value between 0 and 1 indicating the desired probability mass
to include in the \code{yrep} intervals. The default is 0.8.}

\item{group}{A grouping variable of the same length as \code{y}.}
}
\value{
A ggplot object that can be further customized using the
  \pkg{ggplot2} package.
}
\description{
Central interval estimates (quantiles) of \eqn{y^{rep}}{yrep} by time and,
optionally, levels of a grouping variable, with \eqn{y} points overlaid.
}
\examples{
y <- rnorm(50)
yrep <- matrix(rnorm(5000), ncol = 50)
group <- gl(5, 10)
ppc_ts(y, yrep)
ppc_ts_grouped(y, yrep, group)

}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari,
  A., and Rubin, D. B. (2013). \emph{Bayesian Data Analysis.} Chapman & Hall/CRC
  Press, London, third edition. (Ch. 6)
}
\seealso{
\code{\link{set_color_scheme}} to change the aesthetic theme

Other PPCs: \code{\link{distributions}},
  \code{\link{residuals}}, \code{\link{scatterplots}},
  \code{\link{test-statistics}}
}

